refactor(composition): decompose infrastructure composition root
Why:
- Complete Issue #249 while preserving the public composition facade.

What:
- Extract focused configuration, probing, and runtime builder modules.
- Retain compatibility wrappers and synchronize tests and documentation.

How:
- Preserve explicit infrastructure wiring, async behavior, fail-closed probes, and compatibility seams.

Verification:
- : PASS.
- Full WSL quality gate and targeted regression tests: PASS.

Impact:
- Infrastructure composition only; no intended API or live-runtime behavior change.

Limitations:
- No live Docker, Swarm, LXC, Incus, browser, or SonarQube verification was executed.
