+++ src/tiny_swarm_world/infrastructure/composition.py
@@ -1,2863 +1,235 @@
+"""Public composition facade for Tiny Swarm World.
+
+The concrete infrastructure wiring lives in focused internal composition
+modules. This facade deliberately keeps the established import path stable for
+the CLI, integrations, and compatibility tests while avoiding configuration,
+probing, and runtime-construction details here.
+"""
+
 from __future__ import annotations
 
-import asyncio
-import os
-import re
-import shutil
-import subprocess
-import time
-from dataclasses import replace
-from pathlib import Path
-from typing import cast
-from urllib.parse import urlparse
-from uuid import uuid4
+import importlib
+import sys
+from types import FunctionType
+from typing import Any
 
-import requests
+from . import composition_runtime as _runtime
 
-from tiny_swarm_world.application.ports.host import PortHostEnvironmentDetector
-from tiny_swarm_world.application.ports.network import PortWslSocatExposure
-from tiny_swarm_world.application.ports.method_trace import PortMethodTrace
-from tiny_swarm_world.application.services.artifacts import (
-    ArtifactPrepareStep,
-    ArtifactPrepareWorkflow,
-    ArtifactVerifyCheck,
-    ArtifactVerifyWorkflow,
-    ArtifactWorkflowKind,
-    ArtifactReadinessGate,
-    ArtifactWorkflowResult,
-    EnsureContainerImage,
-    EnsureNexusAdminAccess,
-    EnsureNexusDockerHostedRepository,
-    EnsureNexusDockerProxyRepository,
-    EnsureNexusMavenProxyRepository,
-    NexusDockerHostedRepositoryConfiguration,
-    NexusDockerProxyRepositoryConfiguration,
-    NexusMavenProxyRepositoryConfiguration,
-    WaitForNexusReady,
-    StaticArtifactContractPreflight,
-)
-from tiny_swarm_world.application.services.deployment import (
-    DeploymentApplyWorkflow,
-    DeploymentWorkflowKind,
-    DeploymentVerifyWorkflow,
-    EnsureInfisicalSilentInstall,
-    EnsureInfisicalSecretItems,
-    InfisicalSilentInstallConfig,
-    EnsurePortainerEndpoint,
-    EnsurePortainerAdminAccess,
-    EnsureSonarqubeAdminAccess,
-    EnsureSwarmStack,
-    InfisicalSecretItem,
-    InfisicalSecretSyncStep,
-    SecretConsumptionVerifier,
-    SecretDiscoveryStep,
-    SecretEvidenceWriter,
-    SecretManifestRenderer,
-    WriteEffectiveAccessModelEvidence,
-)
-from tiny_swarm_world.application.services.deployment.workflows import (
-    DeploymentApplyStep,
-    DeploymentPreApplyStep,
-)
-from tiny_swarm_world.application.services.deployment.service_stack_plan import (
-    DEFAULT_PORTAINER_ENDPOINT_NAME,
-)
-from tiny_swarm_world.application.ports.progress import PortWorkflowProgress
-from tiny_swarm_world.application.ports.repositories.port_compose_file_repository import (
-    PortComposeFileRepository,
-)
-from tiny_swarm_world.application.ports.ui.port_ui import (
-    AGGREGATE_INSTANCE,
-    STATUS_ERROR,
-    PortUI,
-)
-from tiny_swarm_world.application.services.platform import (
-    AsyncWorkflowStep,
-    LxcDockerInstallService,
-    LxcDockerInstallStep,
-    LxcDockerVerifyStep,
-    LxcProxyDriftRepairService,
-    LxcProxyDriftRepairStep,
-    LxcServiceExposureService,
-    LxcServiceExposureStep,
-    LxcServiceExposureVerifyStep,
-    LxcSwarmBootstrapService,
-    LxcSwarmBootstrapStep,
-    LxcSwarmVerifyStep,
-    NodeProviderDestroyManagedNodesStep,
-    NodeProviderEnsureNodeStep,
-    NodeProviderResetManagedNodesStep,
-    NodeProviderSelectionRequest,
-    NodeProviderSelectionService,
-    NodeProviderVerifyNodeStep,
-    PlatformDestroyWorkflow,
-    PlatformExposeWorkflow,
-    PlatformInitWorkflow,
-    PlatformRepairLxcProxyDriftWorkflow,
-    PlatformReconcileWorkflow,
-    PlatformResetWorkflow,
-    PlatformVerifyWorkflow,
-    PortainerEndpointVerifyStep,
-    PreflightService,
-    SocatManager,
-)
-from tiny_swarm_world.application.services.platform.host import (
-    AuthorizeProjectFilesystem,
-    DetectHostEnvironment,
-    EvaluateProjectFilesystem,
-    HostPreparationAdapterFactory,
-    HostPreparationService,
-)
-from tiny_swarm_world.infrastructure.adapters.file_management.local_file_storage import (
-    LocalFileStorage,
-)
-from tiny_swarm_world.infrastructure.adapters.host import (
-    HostEnvironmentDetector,
-    ProjectFilesystemInspector,
-)
-from tiny_swarm_world.application.services.network import (
-    NetworkDoctorService,
-    NetworkRepairOptions,
-    NetworkRepairService,
-)
-from tiny_swarm_world.application.services.configuration import ConfigurationValidationService
-from tiny_swarm_world.application.services.setup import (
-    SetupWorkflow,
-    SetupWorkflowPhase,
-    SetupWorkflowResult,
-)
-from tiny_swarm_world.domain.artifacts import DEFAULT_CONTAINER_IMAGE_CONTRACTS, ContainerImageContract
-from tiny_swarm_world.domain.deployment import (
-    ServiceStackContract,
-    ServiceStackProfile,
-    service_stack_contracts_for_profile,
-)
-from tiny_swarm_world.domain.inventory import VerificationResult, VerificationStatus
-from tiny_swarm_world.domain.network.port_forwarding_plan import (
-    ForwardingStrategy,
-    PortForwardingPlan,
-)
-from tiny_swarm_world.domain.node_provider import (
-    ManagedLxcBackend,
-    NodeProviderKind,
-    NodeRole,
-    NodeSpec,
-)
-from tiny_swarm_world.domain.preflight import (
-    PreflightResult,
-    LiveConsent,
-    PreflightConfiguration,
-    ProviderPreflightMetadata,
-    default_installation_plan,
-    default_setup_manifest,
-    default_preflight_configuration,
-    default_resource_profiles,
-    RequiredDependency,
-    ResourceThresholds,
-)
-from tiny_swarm_world.infrastructure.adapters.command_runner.command_workflow import CommandWorkflow
-from tiny_swarm_world.infrastructure.adapters.clients.lxc.command.backend_cli import backend_cli
-from tiny_swarm_world.infrastructure.adapters.clients.lxc_node_provider import (
-    AsyncLxcNodeCommandRunner,
-    LxcNodeProvider,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.lxc_container_docker_runtime import (
-    DockerAptMirrorConfiguration,
-    DockerRegistryMirrorConfiguration,
-    LxcContainerDockerRuntime,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.lxc_proxy_device_runtime import (
-    LxcProxyDeviceRuntime,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.lxc.docker.lxc_container_runtime import (
-    LxcContainerRuntime,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.lxc.images.lxc_container_image_publisher import (
-    LxcContainerImagePublisher,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.lxc.services.lxc_nexus_http_client import (
-    LxcNexusHttpClient,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.lxc.services.lxc_portainer_admin_client import (
-    LxcPortainerAdminClient,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.lxc.services.lxc_portainer_http_client import (
-    LxcPortainerHttpClient,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.lxc_swarm_runtime import LxcSwarmRuntime
-from tiny_swarm_world.infrastructure.adapters.clients.infisical_playwright_client import (
-    PlaywrightInfisicalClient,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.infisical_cli_client import (
-    InfisicalCliClient,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.infisical_bootstrap_http_client import (
-    InfisicalBootstrapHttpClient,
-)
-from tiny_swarm_world.infrastructure.adapters.clients.sonarqube_http_client import (
-    SonarqubeHttpClient,
-)
-from tiny_swarm_world.infrastructure.adapters.configuration import (
-    CombinedConfigurationSource,
-    EnvironmentConfigurationSource,
-    ShellEnvFileConfigurationSource,
-)
-from tiny_swarm_world.infrastructure.adapters.ui.progress_trace_ui import (
-    TerminalMethodTrace,
-    TerminalWorkflowProgress,
-)
-from tiny_swarm_world.infrastructure.adapters.ui.factory_ui import FactoryUI
-from tiny_swarm_world.infrastructure.adapters.preflight import (
-    BoundedArtifactReadinessAdapter,
-    DockerManagerReadinessProbe,
-    HostPreflightProbe,
-    HttpArtifactSourceReadiness,
-    HttpEndpointReadinessProbe,
-    LxcProviderPreflightProbe,
-    LocalDirectoryReadinessProbe,
-)
-from tiny_swarm_world.infrastructure.adapters.host.wsl_resource_inspector import WslResourceInspector
-from tiny_swarm_world.infrastructure.adapters.host.hang_diagnostics import ReadOnlyHangDiagnostics
-from tiny_swarm_world.infrastructure.adapters.host.preflight_evidence_writer import PreflightEvidenceWriter
-from tiny_swarm_world.infrastructure.adapters.network import (
-    SubprocessNetworkProbe,
-    SubprocessNetworkRepair,
-    WslSocatExposureAdapter,
-)
-from tiny_swarm_world.infrastructure.adapters.repositories.compose_file_repository_yaml import (
-    ComposeFileRepositoryYaml,
-    resolve_container_image_contracts,
-)
-from tiny_swarm_world.infrastructure.adapters.repositories.node_provider_config_yaml_repository import (
-    NodeProviderConfig,
-    NodeProviderConfigYamlRepository,
-)
-from tiny_swarm_world.infrastructure.adapters.repositories.port_registry_yaml_repository import (
-    PortRegistryYamlRepository,
-)
-from tiny_swarm_world.infrastructure.adapters.repositories.routing_evidence_local_repository import (
-    RoutingEvidenceLocalRepository,
-)
-from tiny_swarm_world.infrastructure.adapters.repositories.project_filesystem_evidence_local_repository import (
-    ProjectFilesystemEvidenceLocalRepository,
-)
-from tiny_swarm_world.infrastructure.os_types import OsTypes
-from tiny_swarm_world.infrastructure.adapters.repositories.verification_evidence_local_repository import (
-    VerificationEvidenceLocalRepository,
-)
-from tiny_swarm_world.infrastructure.logging.logger_factory import LoggerFactory
-from tiny_swarm_world.infrastructure.logging.progress_trace_logging import (
-    CompositeMethodTrace,
-    CompositeWorkflowProgress,
-    LoggingMethodTrace,
-    LoggingWorkflowProgress,
-)
-from tiny_swarm_world.infrastructure.project_paths import ProjectPaths, default_project_paths
-from tiny_swarm_world.infrastructure.process import ProcessRunner, SubprocessProcessRunner
-from tiny_swarm_world.infrastructure.composition_blocked_workflows import (
-    BlockedArtifactWorkflow,
-    BlockedDeploymentWorkflow,
-)
-from tiny_swarm_world.infrastructure.composition_lxc_runtimes import (
-    PrepareLxcStackAssets,
-    ProviderSelectedLxcDockerRuntime,
-    ProviderSelectedLxcProxyDeviceRuntime,
-    ProviderSelectedLxcSwarmRuntime,
-    selected_lxc_backend,
-)
-from tiny_swarm_world.infrastructure.composition_models import (
-    ApplicationServices,
-    ArtifactServices,
-    ArtifactWorkflows,
-    ClusterWorkflows,
-    DeploymentServices,
-    DeploymentWorkflows,
-    PlatformServices,
-    PlatformWorkflows,
-    SetupServices,
-    SetupWorkflows,
+
+_RUNTIME_BUILDER_NAMES = (
+    "build_application_logger",
+    "build_host_environment_detector",
+    "build_host_detection_service",
+    "build_host_preparation_service",
+    "build_project_filesystem_inspector",
+    "build_preflight_service",
+    "build_configuration_validation_service",
+    "build_compose_file_repository",
+    "build_network_doctor_service",
+    "build_read_only_hang_diagnostics",
+    "build_preflight_evidence_writer",
+    "build_process_runner",
+    "build_network_repair_service",
+    "build_network_repair_options",
+    "build_post_install_preflight_service",
+    "build_setup_ui",
+    "run_setup_with_terminal_status",
+    "build_platform_services",
+    "build_artifact_services_for_provider",
+    "build_lxc_artifact_services",
+    "build_deployment_services_for_provider",
+    "build_lxc_deployment_services",
+    "build_setup_services",
+    "build_application_services",
 )
 
-
-DEFAULT_SETUP_SERVICE_PROFILE = ServiceStackProfile.SERVICE_ACCESS
-DEFAULT_OPERATOR_CONFIGURATION_ENV_FILE = Path(".tiny-swarm-world/local/live-installation.env")
-DEFAULT_FIXED_SECRET_ENV_FILE = Path(".tiny-swarm-world/local/fixed-secrets.env")
-_LOCAL_READINESS_SCHEME = "http"
-DEFAULT_PORTAINER_API_URL = "http://localhost:10001"
-PORTAINER_STACK_REQUEST_TIMEOUT_ENVIRONMENT = "TSW_PORTAINER_STACK_REQUEST_TIMEOUT_SECONDS"
-DEFAULT_PORTAINER_STACK_REQUEST_TIMEOUT_SECONDS = 180
-SECRETS_MODE_ENVIRONMENT = "TSW_SECRETS_MODE"
-FIXED_SECRET_ENV_FILE_ENVIRONMENT = "TSW_FIXED_SECRET_ENV_FILE"
-SECRET_MODES = ("generated", "fixed", "infisical")
-SEED_INFISICAL_ITEMS_ENVIRONMENT = "TSW_SEED_INFISICAL_ITEMS"
-INFISICAL_LOGIN_EMAIL_ENVIRONMENT = "TSW_INFISICAL_LOGIN_EMAIL"
-INFISICAL_PASSWORD_ENVIRONMENT = "TSW_INFISICAL_BOOTSTRAP_ADMIN_PASSWORD"
-INFISICAL_READINESS_ATTEMPTS_ENVIRONMENT = "TSW_INFISICAL_READINESS_ATTEMPTS"
-INFISICAL_READINESS_INTERVAL_ENVIRONMENT = "TSW_INFISICAL_READINESS_INTERVAL_SECONDS"
-INFISICAL_URL_ENVIRONMENT = "TSW_INFISICAL_URL"
-INFISICAL_INTERNAL_URL_ENVIRONMENT = "TSW_INFISICAL_INTERNAL_URL"
-INFISICAL_ORGANIZATION_ENVIRONMENT = "TSW_INFISICAL_ORGANIZATION"
-INFISICAL_ADMIN_FIRST_NAME_ENVIRONMENT = "TSW_INFISICAL_ADMIN_FIRST_NAME"
-INFISICAL_ADMIN_LAST_NAME_ENVIRONMENT = "TSW_INFISICAL_ADMIN_LAST_NAME"
-DEFAULT_INFISICAL_ORGANIZATION = "Tiny Swarm World"
-DEFAULT_INFISICAL_READINESS_ATTEMPTS = 720
-DEFAULT_INFISICAL_READINESS_INTERVAL_SECONDS = 5.0
-DEPLOYMENT_VERIFY_TIMEOUT_ENVIRONMENT = "TSW_DEPLOYMENT_VERIFY_TIMEOUT_SECONDS"
-DEFAULT_DEPLOYMENT_VERIFY_TIMEOUT_SECONDS = 300.0
-SWARM_REGISTRY_ENDPOINT_ENVIRONMENT = "TSW_SWARM_REGISTRY_ENDPOINT"
-DEFAULT_SWARM_REGISTRY_ENDPOINT = "127.0.0.1:13500"
-WINDOWS_EXPOSURE_ENVIRONMENT = "TSW_WINDOWS_EXPOSURE"
-LXC_PROXY_LISTEN_ADDRESS_ENVIRONMENT = "TSW_LXC_PROXY_LISTEN_ADDRESS"
-DEFAULT_LXC_PROXY_LISTEN_ADDRESS = "0.0.0.0"
-LXC_UBUNTU_APT_MIRROR_ENVIRONMENT = "TSW_LXC_UBUNTU_APT_MIRROR"
-LXC_UBUNTU_SECURITY_APT_MIRROR_ENVIRONMENT = "TSW_LXC_UBUNTU_SECURITY_APT_MIRROR"
-LXC_DOCKER_APT_MIRROR_ENVIRONMENT = "TSW_LXC_DOCKER_APT_MIRROR"
-LXC_DOCKER_APT_GPG_URL_ENVIRONMENT = "TSW_LXC_DOCKER_APT_GPG_URL"
-NEXUS_DOCKER_HUB_PROXY_REPOSITORY_ENVIRONMENT = "TSW_NEXUS_DOCKER_HUB_PROXY_REPOSITORY"
-NEXUS_DOCKER_HUB_PROXY_PORT_ENVIRONMENT = "TSW_NEXUS_DOCKER_HUB_PROXY_PORT"
-DEFAULT_NEXUS_DOCKER_HUB_PROXY_REPOSITORY = "docker-hub-proxy"
-DEFAULT_NEXUS_DOCKER_HUB_PROXY_PORT = 5001
-DEFAULT_NEXUS_DOCKER_HUB_PROXY_REMOTE_URL = "https://registry-1.docker.io"
-DEFAULT_LXC_MANAGER_PROXY_PROFILE = "docker-swarm-manager"
-NEXUS_IMAGE_ENVIRONMENT = "TSW_NEXUS_IMAGE"
-JENKINS_IMAGE_ENVIRONMENT = "TSW_JENKINS_IMAGE"
-SERVICE_ACCESS_DASHBOARD_IMAGE_ENVIRONMENT = "TSW_SERVICE_ACCESS_DASHBOARD_IMAGE"
-SERVICE_ACCESS_NGINX_IMAGE_ENVIRONMENT = "TSW_SERVICE_ACCESS_NGINX_IMAGE"
-PULSAR_IMAGE_ENVIRONMENT = "TSW_PULSAR_IMAGE"
-PULSAR_MANAGER_IMAGE_ENVIRONMENT = "TSW_PULSAR_MANAGER_IMAGE"
-PULSAR_MANAGER_BOOTSTRAP_IMAGE_ENVIRONMENT = "TSW_PULSAR_MANAGER_BOOTSTRAP_IMAGE"
-TRAEFIK_IMAGE_ENVIRONMENT = "TSW_TRAEFIK_IMAGE"
-DEFAULT_PULSAR_IMAGE = "apachepulsar/pulsar:3.0.17"
-DEFAULT_PULSAR_MANAGER_IMAGE = "apachepulsar/pulsar-manager:v0.4.0"
-TRAEFIK_TLS_CERT_SECRET_NAME_ENVIRONMENT = "TSW_TRAEFIK_TLS_CERT_SECRET_NAME"
-TRAEFIK_TLS_KEY_SECRET_NAME_ENVIRONMENT = "TSW_TRAEFIK_TLS_KEY_SECRET_NAME"
-DEFAULT_TRAEFIK_TLS_CERT_SECRET_NAME = "tsw_traefik_tls_cert"
-DEFAULT_TRAEFIK_TLS_KEY_SECRET_NAME = "tsw_traefik_tls_key"
-INFISICAL_IMAGE_ENVIRONMENT = "TSW_INFISICAL_IMAGE"
-INFISICAL_POSTGRES_IMAGE_ENVIRONMENT = "TSW_INFISICAL_POSTGRES_IMAGE"
-INFISICAL_REDIS_IMAGE_ENVIRONMENT = "TSW_INFISICAL_REDIS_IMAGE"
-INFISICAL_ENCRYPTION_KEY_ENVIRONMENT = "TSW_INFISICAL_ENCRYPTION_KEY"
-INFISICAL_AUTH_SECRET_ENVIRONMENT = "TSW_INFISICAL_AUTH_SECRET"
-INFISICAL_POSTGRES_PASSWORD_ENVIRONMENT = "TSW_INFISICAL_POSTGRES_PASSWORD"
-INFISICAL_REDIS_PASSWORD_ENVIRONMENT = "TSW_INFISICAL_REDIS_PASSWORD"
-REGISTRY_ENDPOINT_PATTERN = re.compile(r"^[A-Za-z0-9][A-Za-z0-9_.-]*(?::\d{1,5})?$")
-DEFAULT_LXC_PLATFORM_NODES = (
-    NodeSpec("swarm-manager", NodeRole.MANAGER, NodeProviderKind.LXC_NATIVE),
-    NodeSpec("swarm-worker-1", NodeRole.WORKER, NodeProviderKind.LXC_NATIVE),
-    NodeSpec("swarm-worker-2", NodeRole.WORKER, NodeProviderKind.LXC_NATIVE),
-)
-LXC_BACKEND_REQUIRED_REASON = "lxc_backend_required"
-_LXC_SUPPORTED_BACKENDS = frozenset((ManagedLxcBackend.INCUS,))
-LXC_BACKEND_REQUIRED_MESSAGE = (
-    "LXC-native workflows require an available or explicitly selected Incus backend."
-)
-_BlockedArtifactWorkflow = BlockedArtifactWorkflow
-_BlockedDeploymentWorkflow = BlockedDeploymentWorkflow
-_ProviderSelectedLxcDockerRuntime = ProviderSelectedLxcDockerRuntime
-_ProviderSelectedLxcSwarmRuntime = ProviderSelectedLxcSwarmRuntime
-_PrepareLxcStackAssets = PrepareLxcStackAssets
-_ProviderSelectedLxcProxyDeviceRuntime = ProviderSelectedLxcProxyDeviceRuntime
-_selected_lxc_backend = selected_lxc_backend
-
-
-class _BlockedPlatformProviderStep:
-    returns_verification_result = True
-
-    def __init__(
-        self,
-        *,
-        target_id: str,
-        provider_request: NodeProviderSelectionRequest,
-        message: str,
-        reason: str,
-    ):
-        self.verification_target_id = target_id
-        self.provider_request = provider_request
-        self.message = message
-        self.reason = reason
-
-    async def run(self) -> VerificationResult:
-        await asyncio.sleep(0)
-        return VerificationResult(
-            target_id=self.verification_target_id,
-            status=VerificationStatus.BLOCKED,
-            message=self.message,
-            evidence={
-                "phase": "pre_apply",
-                "reason": self.reason,
-                "requested_provider": self.provider_request.requested_provider.value,
-            },
-        )
-
-
-class _VerifiedPlatformProviderStep:
-    returns_verification_result = True
-
-    def __init__(
-        self,
-        *,
-        target_id: str,
-        provider_request: NodeProviderSelectionRequest,
-        message: str,
-        reason: str,
-    ):
-        self.verification_target_id = target_id
-        self.provider_request = provider_request
-        self.message = message
-        self.reason = reason
-
-    async def run(self) -> VerificationResult:
-        await asyncio.sleep(0)
-        return VerificationResult(
-            target_id=self.verification_target_id,
-            status=VerificationStatus.VERIFIED,
-            message=self.message,
-            evidence={
-                "phase": "verify",
-                "reason": self.reason,
-                "requested_provider": self.provider_request.requested_provider.value,
-            },
-        )
-
-
-class EndpointReadinessCheck:
-    def __init__(
-        self,
-        service_stack: ServiceStackContract,
-        *,
-        verification_target_id: str | None = None,
-        max_attempts: int = 60,
-        wait_seconds: int = 5,
-        timeout_seconds: int = 5,
-        session: requests.Session | None = None,
-    ) -> None:
-        if max_attempts <= 0:
-            raise ValueError("Endpoint readiness attempts must be positive.")
-        if wait_seconds < 0:
-            raise ValueError("Endpoint readiness wait seconds must not be negative.")
-        if timeout_seconds <= 0:
-            raise ValueError("Endpoint readiness timeout must be positive.")
-        self.service_stack = service_stack
-        self.verification_target_id = verification_target_id or service_stack.service_readiness_target_id
-        self.deployment_target_id = self.verification_target_id
-        self.max_attempts = max_attempts
-        self.wait_seconds = wait_seconds
-        self.timeout_seconds = timeout_seconds
-        self.session = session or requests.Session()
-        self._verification = VerificationResult(
-            target_id=self.verification_target_id,
-            status=VerificationStatus.BLOCKED,
-            message="Endpoint readiness has not run yet.",
-            evidence={
-                "phase": "apply",
-                "reason": "readiness_not_run",
-                "stack_name": service_stack.stack_name,
-            },
-        )
-
-    async def run(self) -> None:
-        self._verification = await self._probe_until_ready_async(phase="apply")
-
-    def verify(self) -> VerificationResult:
-        if self._verification.status is not VerificationStatus.BLOCKED:
-            return self._verification
-        return self._probe_until_ready(phase="verify")
-
-    async def verify_async(self) -> VerificationResult:
-        if self._verification.status is not VerificationStatus.BLOCKED:
-            return self._verification
-        return await self._probe_until_ready_async(phase="verify")
-
-    async def _probe_until_ready_async(self, *, phase: str) -> VerificationResult:
-        if not self.service_stack.endpoints:
-            return VerificationResult(
-                target_id=self.verification_target_id,
-                status=VerificationStatus.BLOCKED,
-                message="Endpoint readiness has no configured endpoints.",
-                evidence=_endpoint_readiness_evidence(self.service_stack, {}, phase=phase, attempt=0),
-            )
-        last_statuses: dict[str, str] = {}
-        for attempt in range(1, self.max_attempts + 1):
-            last_statuses = {}
-            for endpoint in self.service_stack.endpoints:
-                last_statuses[endpoint.name] = await asyncio.to_thread(
-                    _endpoint_status,
-                    self.session,
-                    endpoint.url,
-                    timeout_seconds=self.timeout_seconds,
-                )
-            if all(_endpoint_status_ready(status) for status in last_statuses.values()):
-                return VerificationResult(
-                    target_id=self.verification_target_id,
-                    status=VerificationStatus.VERIFIED,
-                    message="Service endpoints are reachable.",
-                    evidence=_endpoint_readiness_evidence(
-                        self.service_stack,
-                        last_statuses,
-                        phase=phase,
-                        attempt=attempt,
-                    ),
-                )
-            if attempt < self.max_attempts:
-                await asyncio.sleep(self.wait_seconds)
-        return VerificationResult(
-            target_id=self.verification_target_id,
-            status=VerificationStatus.FAILED_TO_VERIFY,
-            message="Service endpoints did not become reachable in time.",
-            evidence=_endpoint_readiness_evidence(
-                self.service_stack,
-                last_statuses,
-                phase=phase,
-                attempt=self.max_attempts,
-            ),
-        )
-
-    def _probe_until_ready(self, *, phase: str) -> VerificationResult:
-        if not self.service_stack.endpoints:
-            return VerificationResult(
-                target_id=self.verification_target_id,
-                status=VerificationStatus.BLOCKED,
-                message="Endpoint readiness has no configured endpoints.",
-                evidence=_endpoint_readiness_evidence(self.service_stack, {}, phase=phase, attempt=0),
-            )
-        last_statuses: dict[str, str] = {}
-        for attempt in range(1, self.max_attempts + 1):
-            last_statuses = {
-                endpoint.name: _endpoint_status(
-                    self.session,
-                    endpoint.url,
-                    timeout_seconds=self.timeout_seconds,
-                )
-                for endpoint in self.service_stack.endpoints
-            }
-            if all(_endpoint_status_ready(status) for status in last_statuses.values()):
-                return VerificationResult(
-                    target_id=self.verification_target_id,
-                    status=VerificationStatus.VERIFIED,
-                    message="Service endpoints are reachable.",
-                    evidence=_endpoint_readiness_evidence(
-                        self.service_stack,
-                        last_statuses,
-                        phase=phase,
-                        attempt=attempt,
-                    ),
-                )
-            if attempt < self.max_attempts:
-                time.sleep(self.wait_seconds)
-        return VerificationResult(
-            target_id=self.verification_target_id,
-            status=VerificationStatus.FAILED_TO_VERIFY,
-            message="Service endpoints did not become reachable in time.",
-            evidence=_endpoint_readiness_evidence(
-                self.service_stack,
-                last_statuses,
-                phase=phase,
-                attempt=self.max_attempts,
-            ),
-        )
-
-
-class _WslSocatExposeStep:
-    returns_verification_result = True
-    verification_target_id = "platform:expose:wsl-socat"
-
-    def __init__(
-        self,
-        socat_manager: SocatManager,
-        socat_exposure: PortWslSocatExposure,
-        *,
-        service_profile: ServiceStackProfile,
-        live_consent: LiveConsent | None,
-        os_type: OsTypes | None = None,
-    ) -> None:
-        self.socat_manager = socat_manager
-        self.socat_exposure = socat_exposure
-        self.service_profile = service_profile
-        self.live_consent = live_consent
-        self.os_type = os_type
-
-    async def run(self) -> VerificationResult:
-        os_type = self.os_type or OsTypes.detect_current()
-        plans = _wsl_socat_forwarding_plans(self.service_profile)
-        commands = self.socat_manager.set_service_socat_ports(os_type, plans)
-        if not commands:
-            return VerificationResult(
-                target_id=self.verification_target_id,
-                status=VerificationStatus.VERIFIED,
-                message="WSL port exposure is not required for this host type.",
-                evidence={
-                    "phase": "verify",
-                    "classification": "not_required",
-                    "os_type": str(getattr(os_type, "value", os_type)),
-                },
-            )
-        if self.live_consent is None or not self.live_consent.accepted:
-            return VerificationResult(
-                target_id=self.verification_target_id,
-                status=VerificationStatus.BLOCKED,
-                message="WSL port exposure requires accepted live infrastructure consent.",
-                evidence={
-                    "phase": "pre_apply",
-                    "classification": "live_mutation_required",
-                    "os_type": str(getattr(os_type, "value", os_type)),
-                    "planned_forward_count": str(len(commands)),
-                },
-            )
-        if not await self.socat_exposure.is_available():
-            return VerificationResult(
-                target_id=self.verification_target_id,
-                status=VerificationStatus.VERIFIED,
-                message="Optional WSL host port forwarding was skipped.",
-                evidence={
-                    "phase": "verify",
-                    "classification": "socat_missing_skipped",
-                    "os_type": str(getattr(os_type, "value", os_type)),
-                    "planned_forward_count": str(len(commands)),
-                    "remediation_hint": "Install the optional WSL forwarding tool and rerun platform expose when Windows-side forwarding is required.",
-                },
-            )
-
-        started_count = 0
-        existing_count = 0
-        failed_count = 0
-        for command in commands:
-            pattern = command.shell_command
-            if await self.socat_exposure.process_exists(pattern):
-                existing_count += 1
-                continue
-            if await self.socat_exposure.start(pattern):
-                started_count += 1
-            else:
-                failed_count += 1
-        status = (
-            VerificationStatus.VERIFIED
-            if failed_count == 0
-            else VerificationStatus.FAILED_TO_APPLY
-        )
-        return VerificationResult(
-            target_id=self.verification_target_id,
-            status=status,
-            message=_wsl_socat_expose_message(status),
-            evidence={
-                "phase": "apply",
-                "classification": (
-                    "wsl_socat_exposed"
-                    if status == VerificationStatus.VERIFIED
-                    else "wsl_socat_expose_failed"
-                ),
-                "os_type": str(getattr(os_type, "value", os_type)),
-                "planned_forward_count": str(len(commands)),
-                "started_count": str(started_count),
-                "existing_count": str(existing_count),
-                "failed_count": str(failed_count),
-            },
-        )
-
-
-def build_application_logger():
-    return LoggerFactory.get_logger("application")
-
-
-def build_host_environment_detector() -> HostEnvironmentDetector:
-    return HostEnvironmentDetector()
-
-
-def build_host_detection_service(
-    detector: PortHostEnvironmentDetector | None = None,
-) -> DetectHostEnvironment:
-    return DetectHostEnvironment(detector or build_host_environment_detector())
-
-
-def build_host_preparation_service(
-    live_consent: LiveConsent | None = None,
-) -> HostPreparationService:
-    project_paths = default_project_paths()
-    bridge_root = project_paths.repository_root / "tools" / "windows"
-    script_path = Path(
-        os.getenv("TSW_WINDOWS_BRIDGE_SCRIPT_PATH", str(bridge_root / "tws-wsl-bridge.ps1"))
-    )
-    config_path = Path(
-        os.getenv("TSW_WINDOWS_BRIDGE_CONFIG_PATH", str(bridge_root / "tws-wsl-bridge.config.json"))
-    )
-    registry_path = Path(
-        os.getenv(
-            "TSW_WINDOWS_BRIDGE_PORT_REGISTRY_PATH",
-            str(project_paths.config_root / "ports.yaml"),
-        )
-    )
-    timeout_seconds = _operator_config_float(
-        "TSW_WINDOWS_BRIDGE_TIMEOUT_SECONDS",
-        120.0,
-        minimum=1.0,
-    )
-    return HostPreparationService(
-        build_host_environment_detector(),
-        HostPreparationAdapterFactory(_build_native_linux_host_preparation),
-        HostPreparationAdapterFactory(
-            lambda: _build_wsl_host_preparation(
-                script_path=script_path,
-                config_path=config_path,
-                registry_path=registry_path,
-                timeout_seconds=timeout_seconds,
-            )
-        ),
-        live_consent,
-    )
-
-
-def _build_native_linux_host_preparation():
-    from tiny_swarm_world.infrastructure.adapters.host import NativeLinuxHostPreparation
-
-    return NativeLinuxHostPreparation()
-
-
-def _build_wsl_host_preparation(
-    *,
-    script_path: Path,
-    config_path: Path,
-    registry_path: Path,
-    timeout_seconds: float,
-):
-    from tiny_swarm_world.infrastructure.adapters.host import (
-        WindowsCommandRunner,
-        WslHostPreparation,
-    )
-
-    return WslHostPreparation(
-        WindowsCommandRunner(),
-        script_path=script_path,
-        config_path=config_path,
-        port_registry_path=registry_path,
-        timeout_seconds=timeout_seconds,
-    )
-
-
-def build_project_filesystem_inspector() -> ProjectFilesystemInspector:
-    return ProjectFilesystemInspector()
-
-
-def _build_project_filesystem_services() -> tuple[
-    EvaluateProjectFilesystem,
-    AuthorizeProjectFilesystem,
-]:
-    inspector = build_project_filesystem_inspector()
-    evidence_repository = ProjectFilesystemEvidenceLocalRepository.from_environment(
-        os.environ,
-        target_inspector=inspector,
-    )
-    return (
-        EvaluateProjectFilesystem(inspector),
-        AuthorizeProjectFilesystem(inspector, evidence_repository),
-    )
-
-
-def build_preflight_service(
-    service_profile: ServiceStackProfile | str = DEFAULT_SETUP_SERVICE_PROFILE,
-    node_provider_request: NodeProviderSelectionRequest | None = None,
-    configuration_validation: ConfigurationValidationService | None = None,
-    allow_wsl_windows_filesystem: bool = False,
-    include_secret_checks: bool = True,
-    include_port_checks: bool = True,
-) -> PreflightService:
-    project_paths = default_project_paths()
-    port_registry = PortRegistryYamlRepository(project_paths=project_paths).load()
-    evaluator, authorizer = _build_project_filesystem_services()
-    return PreflightService(
-        HostPreflightProbe(
-            project_paths=project_paths,
-            host_environment_detector=build_host_environment_detector(),
-            process_runner=build_process_runner(),
-        ),
-        _preflight_configuration_for_provider(service_profile, node_provider_request),
-        configuration_validation=configuration_validation,
-        port_registry=port_registry,
-        project_filesystem_evaluator=evaluator,
-        project_filesystem_authorizer=authorizer,
-        project_path=project_paths.repository_root.as_posix(),
-        allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-        resource_inspector=WslResourceInspector(),
-        evidence_writer=build_preflight_evidence_writer(),
-        artifact_source_readiness=HttpArtifactSourceReadiness(),
-        include_secret_checks=include_secret_checks,
-        include_port_checks=include_port_checks,
-    )
-
-
-def build_configuration_validation_service(
-    env_file: Path | None = None,
-) -> ConfigurationValidationService:
-    resolved_env_file = env_file or Path(
-        os.environ.get("TSW_INSTALL_ENV_FILE", DEFAULT_OPERATOR_CONFIGURATION_ENV_FILE)
-    )
-    return ConfigurationValidationService(
-        CombinedConfigurationSource(
-            (
-                ShellEnvFileConfigurationSource(resolved_env_file),
-                EnvironmentConfigurationSource(),
-            )
-        )
-    )
-
-
-def build_compose_file_repository() -> PortComposeFileRepository:
-    return ComposeFileRepositoryYaml(project_paths=default_project_paths())
-
-
-def build_network_doctor_service() -> NetworkDoctorService:
-    project_paths = default_project_paths()
-    port_registry = PortRegistryYamlRepository(project_paths=project_paths).load()
-    return NetworkDoctorService(
-        SubprocessNetworkProbe(
-            host_environment_detector=build_host_environment_detector()
-        ),
-        port_registry,
-    )
-
-
-def build_read_only_hang_diagnostics() -> ReadOnlyHangDiagnostics:
-    """Build the bounded, non-mutating workflow hang diagnostic adapter."""
-    return ReadOnlyHangDiagnostics(
-        timeout_seconds=_operator_config_float(
-            "TSW_HANG_DIAGNOSTICS_TIMEOUT_SECONDS",
-            10.0,
-            minimum=0.1,
-        )
-    )
-
-
-def build_preflight_evidence_writer() -> PreflightEvidenceWriter:
-    return PreflightEvidenceWriter(default_project_paths().repository_root)
-
-
-def build_process_runner() -> ProcessRunner:
-    """Build the shared bounded infrastructure process runner."""
-    return SubprocessProcessRunner()
-
-
-def build_network_repair_service() -> NetworkRepairService:
-    return NetworkRepairService(
-        SubprocessNetworkProbe(
-            host_environment_detector=build_host_environment_detector()
-        ),
-        SubprocessNetworkRepair(),
-    )
-
-
-def build_network_repair_options(
-    *,
-    runtime: str | None,
-    linux_forwarding: bool,
-    incus: bool,
-    apply: bool,
-) -> NetworkRepairOptions:
-    return NetworkRepairOptions(
-        runtime=runtime,
-        linux_forwarding=linux_forwarding,
-        incus=incus,
-        apply=apply,
-    )
-
-
-def build_post_install_preflight_service(
-    service_profile: ServiceStackProfile | str = DEFAULT_SETUP_SERVICE_PROFILE,
-    node_provider_request: NodeProviderSelectionRequest | None = None,
-    configuration_validation: ConfigurationValidationService | None = None,
-    allow_wsl_windows_filesystem: bool = False,
-) -> PreflightService:
-    project_paths = default_project_paths()
-    configuration = _preflight_configuration_for_provider(service_profile, node_provider_request)
-    evaluator, authorizer = _build_project_filesystem_services()
-    return PreflightService(
-        HostPreflightProbe(
-            project_paths=project_paths,
-            host_environment_detector=build_host_environment_detector(),
-            process_runner=build_process_runner(),
-        ),
-        replace(configuration, required_ports=()),
-        configuration_validation=configuration_validation,
-        project_filesystem_evaluator=evaluator,
-        project_filesystem_authorizer=authorizer,
-        project_path=project_paths.repository_root.as_posix(),
-        allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-        resource_inspector=WslResourceInspector(),
-        evidence_writer=build_preflight_evidence_writer(),
-    )
-
-
-def _build_workflow_progress_sink(ui: PortUI | None = None) -> PortWorkflowProgress:
-    sinks: list[PortWorkflowProgress] = [
-        LoggingWorkflowProgress(LoggerFactory.get_logger("WorkflowProgress"))
-    ]
-    if ui is not None:
-        sinks.append(TerminalWorkflowProgress(ui))
-    return CompositeWorkflowProgress(sinks)
-
-
-def _build_method_trace_sink(ui: PortUI | None = None) -> PortMethodTrace:
-    sinks: list[PortMethodTrace] = [
-        LoggingMethodTrace(LoggerFactory.get_logger("MethodTrace"))
-    ]
-    if ui is not None:
-        sinks.append(TerminalMethodTrace(ui))
-    return CompositeMethodTrace(sinks)
-
-
-def _new_installation_trace_correlation_id() -> str:
-    return f"trace-installation-{uuid4().hex}"
-
-
-def build_setup_ui(*, test_mode: bool = False) -> PortUI:
-    return FactoryUI().get_ui(instances=(), test_mode=test_mode)
-
-
-async def run_setup_with_terminal_status(
-    live_consent: LiveConsent,
-    action: str,
-    service_profile: ServiceStackProfile | str = DEFAULT_SETUP_SERVICE_PROFILE,
-    node_provider_request: NodeProviderSelectionRequest | None = None,
-    allow_wsl_windows_filesystem: bool = False,
-) -> SetupWorkflowResult:
-    if not live_consent.accepted:
-        raise ValueError("setup run requires accepted live consent")
-
-    ui = build_setup_ui()
-    ui.start_in_thread()
-    try:
-        services = build_setup_services(
-            live_consent,
-            service_profile=service_profile,
-            node_provider_request=node_provider_request,
-            allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-            ui=ui,
-            configuration_validation=build_configuration_validation_service(),
-        )
-        match action:
-            case "run":
-                result = await services.workflows.run.run()
-            case _:
-                raise ValueError(f"Unsupported setup workflow action: {action}")
-        ui.update_status(
-            AGGREGATE_INSTANCE,
-            task="setup run",
-            step="finished",
-            result=_setup_result_status(result),
-        )
-        return result
-    except Exception:
-        ui.update_status(
-            AGGREGATE_INSTANCE,
-            task="setup run",
-            step="exception",
-            result=STATUS_ERROR,
-        )
-        raise
-    finally:
-        if ui.ui_thread is not None:
-            await ui.ui_thread
-
-
-def _setup_result_status(result: SetupWorkflowResult) -> str:
-    return result.status.value
-
-
-def build_platform_services(
-    service_profile: ServiceStackProfile | str = DEFAULT_SETUP_SERVICE_PROFILE,
-    live_consent: LiveConsent | None = None,
-    node_provider_request: NodeProviderSelectionRequest | None = None,
-    ui: PortUI | None = None,
-    trace_correlation_id: str | None = None,
-    allow_wsl_windows_filesystem: bool = False,
-) -> PlatformServices:
-    project_paths = default_project_paths()
-    node_provider_config_repository = NodeProviderConfigYamlRepository(
-        project_paths=project_paths
-    )
-    provider_config = node_provider_config_repository.load()
-    provider_request = node_provider_request or _node_provider_request_from_config(provider_config)
-    workflow_progress = _build_workflow_progress_sink(ui)
-    method_trace = _build_method_trace_sink(ui)
-    trace_correlation_id = trace_correlation_id or _new_installation_trace_correlation_id()
-
-    command_workflow = CommandWorkflow()
-    verification_evidence_repository = VerificationEvidenceLocalRepository()
-    preflight = _build_preflight_service_for_request(
-        service_profile,
-        provider_request,
-        project_paths=project_paths,
-        allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-    )
-    post_install_preflight = _build_post_install_preflight_service_for_request(
-        service_profile,
-        provider_request,
-        project_paths=project_paths,
-        allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-    )
-    lxc_runner = AsyncLxcNodeCommandRunner()
-    lxc_node_provider = LxcNodeProvider(
-        config_repository=node_provider_config_repository,
-        runner=lxc_runner,
-        allow_live_mutation=False if live_consent is None else live_consent.accepted,
-        host_resource_inspector=WslResourceInspector(),
-    )
-    node_provider_selection = NodeProviderSelectionService(
-        LxcProviderPreflightProbe(
-            wsl_lxc_capability_available=_wsl_lxc_lifecycle_capability_available,
-        ),
-        lxc_node_provider,
-        lxc_node_provider,
-    )
-    lxc_docker_runtime = _ProviderSelectedLxcDockerRuntime(
-        provider_selection=node_provider_selection,
-        provider_request=provider_request,
-        runner=lxc_runner,
-        allow_live_mutation=False if live_consent is None else live_consent.accepted,
-        registry_mirror_configuration=_lxc_docker_registry_mirror_configuration,
-        apt_mirror_configuration=_lxc_docker_apt_mirror_configuration,
-        docker_runtime_factory=LxcContainerDockerRuntime,
-    )
-    lxc_docker_install = LxcDockerInstallService(lxc_docker_runtime)
-    lxc_docker_verify = LxcDockerInstallService(
-        _ProviderSelectedLxcDockerRuntime(
-            provider_selection=node_provider_selection,
-            provider_request=provider_request,
-            runner=lxc_runner,
-            allow_live_mutation=False,
-            allow_live_inspection=True,
-            registry_mirror_configuration=_lxc_docker_registry_mirror_configuration,
-            apt_mirror_configuration=_lxc_docker_apt_mirror_configuration,
-            docker_runtime_factory=LxcContainerDockerRuntime,
-        )
-    )
-    lxc_swarm_runtime = _ProviderSelectedLxcSwarmRuntime(
-        provider_selection=node_provider_selection,
-        provider_request=provider_request,
-        runner=lxc_runner,
-        allow_live_mutation=False if live_consent is None else live_consent.accepted,
-        proxy_runtime_factory=LxcProxyDeviceRuntime,
-    )
-    lxc_swarm_bootstrap = LxcSwarmBootstrapService(
-        lxc_swarm_runtime,
-        lxc_swarm_runtime,
-    )
-    lxc_swarm_verify = LxcSwarmBootstrapService(
-        _ProviderSelectedLxcSwarmRuntime(
-            provider_selection=node_provider_selection,
-            provider_request=provider_request,
-            runner=lxc_runner,
-            allow_live_mutation=False,
-            allow_live_inspection=True,
-            proxy_runtime_factory=LxcProxyDeviceRuntime,
-        ),
-        lxc_swarm_runtime,
-    )
-    lxc_proxy_runtime = _ProviderSelectedLxcProxyDeviceRuntime(
-        provider_selection=node_provider_selection,
-        provider_request=provider_request,
-        runner=lxc_runner,
-        allow_live_mutation=False if live_consent is None else live_consent.accepted,
-    )
-    lxc_service_exposure = LxcServiceExposureService(
-        lxc_proxy_runtime,
-        gateway_node=_lxc_manager_node(),
-        manager_profile_name=DEFAULT_LXC_MANAGER_PROXY_PROFILE,
-        setup_manifest=default_setup_manifest(service_profile=service_profile),
-        listen_address=_lxc_proxy_listen_address(),
-        logger=LoggerFactory.get_logger("LxcServiceExposureService"),
-    )
-    lxc_service_exposure_verify = LxcServiceExposureService(
-        _ProviderSelectedLxcProxyDeviceRuntime(
-            provider_selection=node_provider_selection,
-            provider_request=provider_request,
-            runner=lxc_runner,
-            allow_live_mutation=False,
-            allow_live_inspection=True,
-        ),
-        gateway_node=_lxc_manager_node(),
-        manager_profile_name=DEFAULT_LXC_MANAGER_PROXY_PROFILE,
-        setup_manifest=default_setup_manifest(service_profile=service_profile),
-        listen_address=_lxc_proxy_listen_address(),
-        logger=LoggerFactory.get_logger("LxcServiceExposureVerifyService"),
-    )
-    lxc_proxy_drift_repair = LxcProxyDriftRepairService(
-        lxc_proxy_runtime,
-        gateway_node=_lxc_manager_node(),
-        manager_profile_name=DEFAULT_LXC_MANAGER_PROXY_PROFILE,
-        setup_manifest=default_setup_manifest(service_profile=service_profile),
-        listen_address=_lxc_proxy_listen_address(),
-    )
-    socat_manager = SocatManager()
-    socat_exposure = WslSocatExposureAdapter()
-    init_steps = _platform_init_steps(
-        provider_request=provider_request,
-        node_provider_selection=node_provider_selection,
-    )
-    workflows = PlatformWorkflows(
-        init=PlatformInitWorkflow(
-            init_steps,
-            verification_evidence_repository=verification_evidence_repository,
-            pre_apply_guard=(
-                SetupWorkflowPhase(
-                    "platform init preflight",
-                    lambda: _platform_init_pre_apply_guard(
-                        preflight,
-                        node_provider_selection,
-                        provider_request,
-                        live_consent,
-                    ),
-                    method_trace=method_trace,
-                    trace_correlation_id=trace_correlation_id,
-                )
-                if live_consent is not None
-                else None
-            ),
-            progress=workflow_progress,
-            method_trace=method_trace,
-            trace_correlation_id=trace_correlation_id,
-        ),
-        reconcile=PlatformReconcileWorkflow(
-            _platform_reconcile_steps(
-                provider_request=provider_request,
-                node_provider_selection=node_provider_selection,
-            ),
-            verification_evidence_repository=verification_evidence_repository,
-            progress=workflow_progress,
-            method_trace=method_trace,
-            trace_correlation_id=trace_correlation_id,
-        ),
-        expose=PlatformExposeWorkflow(
-            _platform_expose_steps(
-                lxc_service_exposure,
-                socat_manager,
-                socat_exposure,
-                service_profile=ServiceStackProfile(service_profile),
-                live_consent=live_consent,
-            ),
-            verification_evidence_repository=verification_evidence_repository,
-            progress=workflow_progress,
-            method_trace=method_trace,
-            trace_correlation_id=trace_correlation_id,
-        ),
-        repair_lxc_proxy_drift=PlatformRepairLxcProxyDriftWorkflow(
-            _platform_repair_lxc_proxy_drift_steps(lxc_proxy_drift_repair),
-            verification_evidence_repository=verification_evidence_repository,
-            progress=workflow_progress,
-            method_trace=method_trace,
-            trace_correlation_id=trace_correlation_id,
-        ),
-        reset=PlatformResetWorkflow(
-            _platform_reset_steps(
-                provider_request,
-                node_provider_selection,
-            ),
-            verification_evidence_repository=verification_evidence_repository,
-            progress=workflow_progress,
-            method_trace=method_trace,
-            trace_correlation_id=trace_correlation_id,
-        ),
-        destroy=PlatformDestroyWorkflow(
-            _platform_destroy_steps(
-                provider_request,
-                node_provider_selection,
-            ),
-            verification_evidence_repository=verification_evidence_repository,
-            progress=workflow_progress,
-            method_trace=method_trace,
-            trace_correlation_id=trace_correlation_id,
-        ),
-        verify=PlatformVerifyWorkflow(
-            _platform_verify_steps(
-                post_install_preflight,
-                provider_request=provider_request,
-                node_provider_selection=node_provider_selection,
-                lxc_service_exposure=lxc_service_exposure_verify,
-            ),
-            progress=workflow_progress,
-            method_trace=method_trace,
-            trace_correlation_id=trace_correlation_id,
-            verify_retry_attempts=6,
-            verify_retry_delay_seconds=10.0,
-        ),
-        cluster=ClusterWorkflows(
-            docker=PlatformInitWorkflow(
-                _cluster_docker_steps(lxc_docker_install),
-                verification_evidence_repository=verification_evidence_repository,
-                progress=workflow_progress,
-                method_trace=method_trace,
-                trace_correlation_id=trace_correlation_id,
-            ),
-            swarm_bootstrap=PlatformInitWorkflow(
-                _cluster_swarm_bootstrap_steps(lxc_swarm_bootstrap),
-                verification_evidence_repository=verification_evidence_repository,
-                progress=workflow_progress,
-                method_trace=method_trace,
-                trace_correlation_id=trace_correlation_id,
-            ),
-            verify=PlatformVerifyWorkflow(
-                _cluster_verify_steps(
-                    lxc_docker_verify,
-                    lxc_swarm_verify,
-                ),
-                progress=workflow_progress,
-                method_trace=method_trace,
-                trace_correlation_id=trace_correlation_id,
-                verify_retry_attempts=6,
-                verify_retry_delay_seconds=10.0,
-            ),
-        ),
-    )
-
-    return PlatformServices(
-        command_workflow=command_workflow,
-        lxc_docker_install=lxc_docker_install,
-        lxc_proxy_drift_repair=lxc_proxy_drift_repair,
-        lxc_service_exposure=lxc_service_exposure,
-        lxc_swarm_bootstrap=lxc_swarm_bootstrap,
-        preflight=preflight,
-        lxc_node_provider=lxc_node_provider,
-        node_provider_selection=node_provider_selection,
-        socat_manager=socat_manager,
-        workflows=workflows,
-    )
-
-
-def build_artifact_services_for_provider(
-    node_provider_request: NodeProviderSelectionRequest | None = None,
-    ui: PortUI | None = None,
-    progress: PortWorkflowProgress | None = None,
-) -> ArtifactServices:
-    provider_request = node_provider_request or _default_node_provider_request()
-    backend = _lxc_backend_for_provider_request(provider_request)
-    if backend is not None:
-        return build_lxc_artifact_services(backend=backend, ui=ui, progress=progress)
-    return ArtifactServices(
-        workflows=ArtifactWorkflows(
-            prepare=cast(
-                ArtifactPrepareWorkflow,
-                _BlockedArtifactWorkflow(
-                    ArtifactWorkflowKind.PREPARE,
-                    LXC_BACKEND_REQUIRED_REASON,
-                ),
-            ),
-            verify=cast(
-                ArtifactVerifyWorkflow,
-                _BlockedArtifactWorkflow(
-                    ArtifactWorkflowKind.VERIFY,
-                    LXC_BACKEND_REQUIRED_REASON,
-                ),
-            ),
-        )
-    )
-
-
-def build_lxc_artifact_services(
-    *,
-    backend: ManagedLxcBackend,
-    ui: PortUI | None = None,
-    progress: PortWorkflowProgress | None = None,
-) -> ArtifactServices:
-    project_paths = default_project_paths()
-    process_runner = build_process_runner()
-    nexus_admin_password = _operator_secret_value("TSW_NEXUS_ADMIN_PASSWORD")
-    nexus_client = LxcNexusHttpClient(backend=backend)
-    container_runtime = LxcContainerRuntime(
-        backend=backend,
-        process_runner=process_runner,
-        node_names=tuple(node.name for node in DEFAULT_LXC_PLATFORM_NODES),
-    )
-    image_publisher = LxcContainerImagePublisher(
-        backend=backend,
-        process_runner=process_runner,
-        registry_username="admin",
-        registry_password=nexus_admin_password,
-        project_paths=project_paths,
-    )
-    wait_for_nexus_ready = WaitForNexusReady(
-        nexus_client=nexus_client,
-        max_attempts=60,
-        wait_seconds=10,
-        progress=progress,
-    )
-    ensure_nexus_admin_access = EnsureNexusAdminAccess(
-        nexus_client=nexus_client,
-        container_runtime=container_runtime,
-        admin_username="admin",
-        admin_password=nexus_admin_password,
-        container_name_filter="nexus",
-        initial_password_path="/nexus-data/admin.password",
-        max_attempts=60,
-        wait_seconds=10,
-        ui=ui,
-        progress=progress,
-    )
-    nexus_repository_steps = (
-        EnsureNexusDockerHostedRepository(
-            nexus_client=nexus_client,
-            configuration=NexusDockerHostedRepositoryConfiguration(
-                repository_name="docker-hosted",
-                http_port=5000,
-                admin_username="admin",
-                admin_password=nexus_admin_password,
-            ),
-        ),
-        EnsureNexusDockerProxyRepository(
-            nexus_client=nexus_client,
-            configuration=NexusDockerProxyRepositoryConfiguration(
-                repository_name=_nexus_docker_hub_proxy_repository_name(),
-                http_port=_nexus_docker_hub_proxy_port(),
-                remote_url=_nexus_docker_proxy_remote_url(),
-                admin_username="admin",
-                admin_password=nexus_admin_password,
-            ),
-        ),
-        EnsureNexusMavenProxyRepository(
-            nexus_client=nexus_client,
-            configuration=NexusMavenProxyRepositoryConfiguration(
-                repository_name="maven-central-proxy",
-                remote_url="https://repo1.maven.org/maven2/",
-                admin_username="admin",
-                admin_password=nexus_admin_password,
-            ),
-        ),
-    )
-    image_steps = tuple(
-        EnsureContainerImage(image_publisher, contract)
-        for contract in _container_image_contracts_from_environment()
-    )
-    bootstrap_steps = (
-        wait_for_nexus_ready,
-        ensure_nexus_admin_access,
-        *nexus_repository_steps,
-    )
-    checks = cast(
-        tuple[ArtifactPrepareStep, ...],
-        (
-            *bootstrap_steps,
-            *image_steps,
-        ),
-    )
-    verify_checks = cast(tuple[ArtifactVerifyCheck, ...], checks)
-    return ArtifactServices(
-        workflows=ArtifactWorkflows(
-            prepare=ArtifactPrepareWorkflow(checks, bootstrap_steps=bootstrap_steps),
-            verify=ArtifactVerifyWorkflow(verify_checks),
-        )
-    )
-
-
-def build_deployment_services_for_provider(
-    service_profile: ServiceStackProfile | str = DEFAULT_SETUP_SERVICE_PROFILE,
-    node_provider_request: NodeProviderSelectionRequest | None = None,
-    ui: PortUI | None = None,
-    progress: PortWorkflowProgress | None = None,
-) -> DeploymentServices:
-    provider_request = node_provider_request or _default_node_provider_request()
-    backend = _lxc_backend_for_provider_request(provider_request)
-    backend_executable = None if backend is None else backend_cli(backend)
-    if backend is not None and backend in _LXC_SUPPORTED_BACKENDS and backend_executable is not None and shutil.which(backend_executable):
-        return build_lxc_deployment_services(
-            service_profile=service_profile,
-            backend=backend,
-            ui=ui,
-            progress=progress,
-        )
-    return DeploymentServices(
-        workflows=DeploymentWorkflows(
-            bootstrap=cast(
-                DeploymentApplyWorkflow,
-                _BlockedDeploymentWorkflow(
-                    DeploymentWorkflowKind.BOOTSTRAP,
-                    LXC_BACKEND_REQUIRED_REASON,
-                ),
-            ),
-            apply=cast(
-                DeploymentApplyWorkflow,
-                _BlockedDeploymentWorkflow(
-                    DeploymentWorkflowKind.APPLY,
-                    LXC_BACKEND_REQUIRED_REASON,
-                ),
-            ),
-            verify=cast(
-                DeploymentVerifyWorkflow,
-                _BlockedDeploymentWorkflow(
-                    DeploymentWorkflowKind.VERIFY,
-                    LXC_BACKEND_REQUIRED_REASON,
-                ),
-            ),
-        )
-    )
-
-
-def build_lxc_deployment_services(
-    *,
-    backend: ManagedLxcBackend,
-    service_profile: ServiceStackProfile | str = DEFAULT_SETUP_SERVICE_PROFILE,
-    ui: PortUI | None = None,
-    progress: PortWorkflowProgress | None = None,
-) -> DeploymentServices:
-    project_paths = default_project_paths()
-    local_file_storage = LocalFileStorage()
-    selected_service_profile = ServiceStackProfile(service_profile)
-    service_stack_contracts = service_stack_contracts_for_profile(selected_service_profile)
-    compose_repository = ComposeFileRepositoryYaml(
-        project_paths=project_paths,
-        service_profile=selected_service_profile,
-    )
-    routing_evidence_step = WriteEffectiveAccessModelEvidence(
-        effective_access_model_repository=compose_repository,
-        routing_evidence_repository=RoutingEvidenceLocalRepository(
-            project_paths=project_paths,
-        ),
-        service_profile=selected_service_profile,
-    )
-    traefik_tls_cert_secret_name = _operator_config_value(
-        TRAEFIK_TLS_CERT_SECRET_NAME_ENVIRONMENT,
-        DEFAULT_TRAEFIK_TLS_CERT_SECRET_NAME,
-    )
-    traefik_tls_key_secret_name = _operator_config_value(
-        TRAEFIK_TLS_KEY_SECRET_NAME_ENVIRONMENT,
-        DEFAULT_TRAEFIK_TLS_KEY_SECRET_NAME,
-    )
-    swarm_runtime = LxcSwarmRuntime(
-        backend=backend,
-        process_runner=build_process_runner(),
-        project_paths=project_paths,
-        service_access_dashboard_renderer=compose_repository.render_service_access_dashboard,
-        traefik_tls_cert_secret_name=traefik_tls_cert_secret_name,
-        traefik_tls_key_secret_name=traefik_tls_key_secret_name,
-    )
-    stack_environment = _deployment_stack_environment(selected_service_profile)
-    secret_manifest_entries = SecretManifestRenderer(local_file_storage).run()
-    portainer_admin_client = LxcPortainerAdminClient(backend=backend)
-    portainer_client = LxcPortainerHttpClient(
-        backend=backend,
-        username="admin",
-        password=_operator_secret_value("TSW_PORTAINER_ADMIN_PASSWORD"),
-        stack_request_timeout_seconds=_operator_config_int(
-            PORTAINER_STACK_REQUEST_TIMEOUT_ENVIRONMENT,
-            DEFAULT_PORTAINER_STACK_REQUEST_TIMEOUT_SECONDS,
-            minimum=1,
-        ),
-    )
-    stack_steps = {
-        contract.stack_name: EnsureSwarmStack(
-            compose_repository=compose_repository,
-            swarm_runtime=swarm_runtime,
-            service_stack=contract,
-            stack_environment=stack_environment.get(contract.stack_name),
-        )
-        for contract in service_stack_contracts
-    }
-    bootstrap_steps = (
-        stack_steps["portainer"],
-        EnsurePortainerAdminAccess(
-            portainer_admin_client=portainer_admin_client,
-            username="admin",
-            password=_operator_secret_value("TSW_PORTAINER_ADMIN_PASSWORD"),
-            max_attempts=60,
-            wait_seconds=5,
-            ui=ui,
-        ),
-        EnsurePortainerEndpoint(
-            portainer_client=portainer_client,
-            endpoint_name=DEFAULT_PORTAINER_ENDPOINT_NAME,
-        ),
-        stack_steps["nexus"],
-    )
-    application_steps: tuple[object, ...] = tuple(
-        stack_steps[contract.stack_name]
-        for contract in service_stack_contracts
-        if contract.stack_name not in {"portainer", "nexus", "traefik"}
-    )
-    if selected_service_profile is ServiceStackProfile.SERVICE_ACCESS:
-        application_steps = _prioritize_infisical_apply_steps(
-            (stack_steps["traefik"], *application_steps)
-        )
-    sonarqube_admin_step = EnsureSonarqubeAdminAccess(
-        sonarqube_client=SonarqubeHttpClient(
-            _operator_config_value(
-                "TSW_SONARQUBE_URL",
-                _local_http_url("localhost", "12000"),
-            )
-        ),
-        username=_operator_config_value("TSW_SONARQUBE_ADMIN_USERNAME", "admin"),
-        password=_operator_secret_value("TSW_SONARQUBE_ADMIN_PASSWORD"),
-        progress=progress,
-    )
-    application_steps = _with_post_stack_steps(
-        application_steps,
-        "sonarqube",
-        (sonarqube_admin_step,),
-    )
-    infisical_cli_client = InfisicalCliClient()
-    service_stack_by_name = {contract.stack_name: contract for contract in service_stack_contracts}
-    infisical_apply_readiness_steps = _infisical_apply_readiness_steps(
-        selected_service_profile,
-        service_stack_by_name=service_stack_by_name,
-    )
-    infisical_bootstrap_steps = _infisical_bootstrap_steps(
-        selected_service_profile,
-        cli=infisical_cli_client,
-        swarm_runtime=swarm_runtime,
-    )
-    secret_discovery_step = SecretDiscoveryStep(
-        storage=local_file_storage,
-        manifest_entries=secret_manifest_entries,
-    )
-    infisical_secret_sync_step = InfisicalSecretSyncStep(
-        cli=infisical_cli_client,
-        storage=local_file_storage,
-        manifest_entries=secret_manifest_entries,
-        fixed_env_file=_fixed_secret_env_file(),
-        mode=_secret_mode(),  # type: ignore[arg-type]
-        process_environment=os.environ,
-    )
-    secret_consumption_step = SecretConsumptionVerifier(
-        manifest_entries=secret_manifest_entries,
-        stack_environment=stack_environment,
-        non_stack_consumer_refs={
-            "TSW_NEXUS_ADMIN_PASSWORD": EnsureNexusAdminAccess.verification_target_id,
-            "TSW_PORTAINER_ADMIN_PASSWORD": EnsurePortainerAdminAccess.verification_target_id,
-            "TSW_SONARQUBE_ADMIN_PASSWORD": EnsureSonarqubeAdminAccess.verification_target_id,
-        },
-    )
-    secret_evidence_step = SecretEvidenceWriter(
-        storage=local_file_storage,
-        discovery=secret_discovery_step,
-        sync=infisical_secret_sync_step,
-        consumption=secret_consumption_step,
-    )
-    infisical_secret_management_steps = (
-        secret_discovery_step,
-        *infisical_bootstrap_steps,
-        infisical_secret_sync_step,
-        secret_consumption_step,
-        secret_evidence_step,
-    )
-    infisical_seed_steps = _infisical_secret_seed_steps(selected_service_profile)
-    readiness_checks = tuple(
-        EndpointReadinessCheck(
-            service_stack=contract,
-            max_attempts=60,
-            wait_seconds=5,
-        )
-        for contract in service_stack_contracts
-    )
-    pre_apply_steps: list[DeploymentPreApplyStep] = [
-        routing_evidence_step,
-        _PrepareLxcStackAssets(swarm_runtime, "traefik"),
-        _PrepareLxcStackAssets(swarm_runtime, "swagger"),
-    ]
-    if selected_service_profile is ServiceStackProfile.SERVICE_ACCESS:
-        pre_apply_steps.append(_PrepareLxcStackAssets(swarm_runtime, "service-access"))
-
-    return DeploymentServices(
-        workflows=DeploymentWorkflows(
-            bootstrap=DeploymentApplyWorkflow(
-                bootstrap_steps,
-                kind=DeploymentWorkflowKind.BOOTSTRAP,
-            ),
-            apply=DeploymentApplyWorkflow(
-                cast(
-                    tuple[DeploymentApplyStep, ...],
-                    _with_infisical_post_apply_steps(
-                        application_steps,
-                        (
-                            *infisical_apply_readiness_steps,
-                            *infisical_secret_management_steps,
-                            *infisical_seed_steps,
-                        ),
-                    ),
-                ),
-                pre_apply_steps=tuple(pre_apply_steps),
-            ),
-            verify=DeploymentVerifyWorkflow(
-                readiness_checks,
-                timeout_seconds=_operator_config_float(
-                    DEPLOYMENT_VERIFY_TIMEOUT_ENVIRONMENT,
-                    DEFAULT_DEPLOYMENT_VERIFY_TIMEOUT_SECONDS,
-                    minimum=1.0,
-                ),
-            ),
-        )
-    )
-
-
-def _build_artifact_readiness_gate(project_paths: ProjectPaths) -> ArtifactReadinessGate:
-    nexus_base_url = os.getenv(
-        "TSW_NEXUS_READINESS_BASE_URL",
-        f"{_LOCAL_READINESS_SCHEME}://127.0.0.1:13081",
-    ).rstrip("/")
-    registry_base_url = _http_readiness_base_url(_swarm_registry_endpoint())
-    manager_storage_path = Path(
-        os.getenv("TSW_MANAGER_STORAGE_PATH", "/var/lib/docker")
-    )
-    return ArtifactReadinessGate(
-        BoundedArtifactReadinessAdapter(
-            {
-                "docker:manager": DockerManagerReadinessProbe(),
-                "registry:endpoint": HttpEndpointReadinessProbe(
-                    f"{registry_base_url}/v2/",
-                    probe_kind="registry_endpoint",
-                ),
-                "nexus:endpoint": HttpEndpointReadinessProbe(
-                    f"{nexus_base_url}/service/rest/v1/status",
-                    probe_kind="nexus_endpoint",
-                ),
-                "nexus:repositories": HttpEndpointReadinessProbe(
-                    f"{nexus_base_url}/service/rest/v1/repositories",
-                    probe_kind="nexus_repositories",
-                ),
-                "storage:manager": LocalDirectoryReadinessProbe(
-                    manager_storage_path,
-                    probe_kind="manager_storage",
-                ),
-                "build:inputs": LocalDirectoryReadinessProbe(
-                    project_paths.repository_root,
-                    probe_kind="build_inputs",
-                ),
-                "pull:public": HttpEndpointReadinessProbe(
-                    os.getenv(
-                        "TSW_PUBLIC_PULL_READINESS_URL",
-                        "https://registry-1.docker.io/v2/",
-                    ),
-                    probe_kind="public_pull",
-                ),
-            }
-        )
-    )
-
-
-def _http_readiness_base_url(endpoint: str) -> str:
-    normalized = endpoint.strip()
-    parsed = urlparse(normalized)
-    if parsed.scheme in {"http", "https"}:
-        return normalized.rstrip("/")
-    return f"{_LOCAL_READINESS_SCHEME}://{normalized.rstrip('/')}"
-
-
-def build_setup_services(
-    live_consent: LiveConsent,
-    service_profile: ServiceStackProfile | str = DEFAULT_SETUP_SERVICE_PROFILE,
-    node_provider_request: NodeProviderSelectionRequest | None = None,
-    ui: PortUI | None = None,
-    configuration_validation: ConfigurationValidationService | None = None,
-    allow_wsl_windows_filesystem: bool = False,
-) -> SetupServices:
-    project_paths = default_project_paths()
-    selected_service_profile = ServiceStackProfile(service_profile)
-    preflight = _build_preflight_service_for_request(
-        selected_service_profile,
-        node_provider_request,
-        configuration_validation=configuration_validation,
-        project_paths=project_paths,
-        allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-    )
-    artifact_contract_preflight = StaticArtifactContractPreflight(
-        compose_repository=ComposeFileRepositoryYaml(
-            project_paths=project_paths,
-            service_profile=selected_service_profile,
-        ),
-        storage=LocalFileStorage(),
-    )
-    artifact_readiness_gate = _build_artifact_readiness_gate(project_paths)
-    host_preparation = build_host_preparation_service(live_consent)
-    trace_correlation_id = _new_installation_trace_correlation_id()
-    workflow_progress = _build_workflow_progress_sink(ui)
-    platform = _build_platform_services_for_request(
-        service_profile,
-        live_consent,
-        node_provider_request,
-        ui=ui,
-        trace_correlation_id=trace_correlation_id,
-        allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-    )
-    artifacts = build_artifact_services_for_provider(
-        node_provider_request=node_provider_request,
-        ui=ui,
-        progress=workflow_progress,
-    )
-    deployment = _build_deployment_services_for_request(
-        service_profile=service_profile,
-        node_provider_request=node_provider_request,
-        ui=ui,
-        progress=workflow_progress,
-    )
-    method_trace = _build_method_trace_sink(ui)
-    static_preflight_result: PreflightResult | None = None
-    artifact_bootstrap_result: ArtifactWorkflowResult | None = None
-
-    def run_artifact_contract_preflight() -> PreflightResult:
-        nonlocal static_preflight_result
-        static_preflight_result = artifact_contract_preflight.run()
-        return static_preflight_result
-
-    async def run_artifact_bootstrap():
-        nonlocal artifact_bootstrap_result
-        artifact_bootstrap_result = await artifacts.workflows.prepare.run_bootstrap()
-        return artifact_bootstrap_result
-
-    def run_artifact_readiness_gate() -> PreflightResult:
-        return artifact_readiness_gate.run(
-            static_preflight=static_preflight_result,
-            artifact_bootstrap=artifact_bootstrap_result,
-        )
-
-    def traced_phase(name: str, runner) -> SetupWorkflowPhase:
-        return SetupWorkflowPhase(
-            name,
-            runner,
-            method_trace=method_trace,
-            trace_correlation_id=trace_correlation_id,
-        )
-
-    return SetupServices(
-        workflows=SetupWorkflows(
-            run=SetupWorkflow(
-                (
-                    traced_phase("preflight", lambda: preflight.run(live_consent)),
-                    traced_phase(
-                        "artifact contract preflight",
-                        run_artifact_contract_preflight,
-                    ),
-                    traced_phase("host prepare", host_preparation.prepare),
-                    traced_phase("host verify", host_preparation.verify),
-                    traced_phase("platform init", lambda: platform.workflows.init.run()),
-                    traced_phase(
-                        "platform reconcile",
-                        lambda: platform.workflows.reconcile.run(),
-                    ),
-                    traced_phase(
-                        "cluster docker",
-                        lambda: platform.workflows.cluster.docker.run(),
-                    ),
-                    traced_phase(
-                        "cluster swarm bootstrap",
-                        lambda: platform.workflows.cluster.swarm_bootstrap.run(),
-                    ),
-                    traced_phase(
-                        "cluster verify",
-                        lambda: platform.workflows.cluster.verify.run(),
-                    ),
-                    traced_phase(
-                        "platform expose",
-                        lambda: platform.workflows.expose.run(),
-                    ),
-                    traced_phase(
-                        "deployment bootstrap",
-                        lambda: deployment.workflows.bootstrap.run(),
-                    ),
-                    traced_phase("artifact bootstrap", run_artifact_bootstrap),
-                    traced_phase("artifact readiness gate", run_artifact_readiness_gate),
-                    traced_phase(
-                        "artifacts prepare",
-                        lambda: artifacts.workflows.prepare.run_after_bootstrap(
-                            artifact_bootstrap_result
-                        ),
-                    ),
-                    traced_phase(
-                        "artifacts verify",
-                        lambda: artifacts.workflows.verify.run(),
-                    ),
-                    traced_phase(
-                        "deployment apply",
-                        lambda: deployment.workflows.apply.run(),
-                    ),
-                    traced_phase(
-                        "deployment verify",
-                        lambda: deployment.workflows.verify.run(),
-                    ),
-                    traced_phase("platform verify", lambda: platform.workflows.verify.run()),
-                ),
-                live_consent=live_consent,
-                progress=workflow_progress,
-                method_trace=method_trace,
-                trace_correlation_id=trace_correlation_id,
-                installation_plan=default_installation_plan(),
-                timeout_seconds=_operator_config_float(
-                    "TSW_SETUP_WORKFLOW_TIMEOUT_SECONDS",
-                    3600.0,
-                    minimum=1.0,
-                ),
-                heartbeat_interval_seconds=_operator_config_float(
-                    "TSW_SETUP_HEARTBEAT_INTERVAL_SECONDS",
-                    30.0,
-                    minimum=1.0,
-                ),
-                max_concurrency=_operator_config_int(
-                    "TSW_SETUP_MAX_CONCURRENCY",
-                    2,
-                    minimum=1,
-                ),
-            )
-        )
-    )
-
-
-def build_application_services(
-    live_consent: LiveConsent | None = None,
-    service_profile: ServiceStackProfile | str = DEFAULT_SETUP_SERVICE_PROFILE,
-    node_provider_request: NodeProviderSelectionRequest | None = None,
-    ui: PortUI | None = None,
-    allow_wsl_windows_filesystem: bool = False,
-) -> ApplicationServices:
-    return ApplicationServices(
-        platform=_build_platform_services_for_request(
-            service_profile,
-            live_consent,
-            node_provider_request,
-            ui=ui,
-            allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-        ),
-        artifacts=build_artifact_services_for_provider(
-            node_provider_request=node_provider_request
-        ),
-        deployment=_build_deployment_services_for_request(
-            service_profile=service_profile,
-            node_provider_request=node_provider_request,
-            ui=ui,
-        ),
-    )
-
-
-def _build_platform_services_for_request(
-    service_profile: ServiceStackProfile | str,
-    live_consent: LiveConsent | None,
-    node_provider_request: NodeProviderSelectionRequest | None,
-    ui: PortUI | None = None,
-    trace_correlation_id: str | None = None,
-    allow_wsl_windows_filesystem: bool = False,
-) -> PlatformServices:
-    if node_provider_request is None:
-        if ui is None and trace_correlation_id is None:
-            return build_platform_services(
-                service_profile=service_profile,
-                live_consent=live_consent,
-                allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-            )
-        return build_platform_services(
-            service_profile=service_profile,
-            live_consent=live_consent,
-            ui=ui,
-            trace_correlation_id=trace_correlation_id,
-            allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-        )
-    if ui is None and trace_correlation_id is None:
-        return build_platform_services(
-            service_profile=service_profile,
-            live_consent=live_consent,
-            node_provider_request=node_provider_request,
-            allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-        )
-    return build_platform_services(
-        service_profile=service_profile,
-        live_consent=live_consent,
-        node_provider_request=node_provider_request,
-        ui=ui,
-        trace_correlation_id=trace_correlation_id,
-        allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-    )
-
-
-def _build_deployment_services_for_request(
-    service_profile: ServiceStackProfile | str,
-    node_provider_request: NodeProviderSelectionRequest | None,
-    ui: PortUI | None = None,
-    progress: PortWorkflowProgress | None = None,
-) -> DeploymentServices:
-    if ui is None:
-        if progress is None:
-            return build_deployment_services_for_provider(
-                service_profile=service_profile,
-                node_provider_request=node_provider_request,
-            )
-        return build_deployment_services_for_provider(
-            service_profile=service_profile,
-            node_provider_request=node_provider_request,
-            progress=progress,
-        )
-    if progress is None:
-        return build_deployment_services_for_provider(
-            service_profile=service_profile,
-            node_provider_request=node_provider_request,
-            ui=ui,
-        )
-    return build_deployment_services_for_provider(
-        service_profile=service_profile,
-        node_provider_request=node_provider_request,
-        ui=ui,
-        progress=progress,
-    )
-
-
-def _build_preflight_service_for_request(
-    service_profile: ServiceStackProfile | str,
-    node_provider_request: NodeProviderSelectionRequest | None,
-    configuration_validation: ConfigurationValidationService | None = None,
-    project_paths: ProjectPaths | None = None,
-    allow_wsl_windows_filesystem: bool = False,
-) -> PreflightService:
-    paths = project_paths or default_project_paths()
-    if node_provider_request is None:
-        node_provider_request = None
-    port_registry = PortRegistryYamlRepository(project_paths=paths).load()
-    evaluator, authorizer = _build_project_filesystem_services()
-    return PreflightService(
-        HostPreflightProbe(
-            project_paths=paths,
-            host_environment_detector=build_host_environment_detector(),
-            process_runner=build_process_runner(),
-        ),
-        _preflight_configuration_for_provider(service_profile, node_provider_request),
-        configuration_validation=configuration_validation,
-        port_registry=port_registry,
-        project_filesystem_evaluator=evaluator,
-        project_filesystem_authorizer=authorizer,
-        project_path=paths.repository_root.as_posix(),
-        allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-        resource_inspector=WslResourceInspector(),
-        evidence_writer=build_preflight_evidence_writer(),
-        artifact_source_readiness=HttpArtifactSourceReadiness(),
-    )
-
-
-def _build_post_install_preflight_service_for_request(
-    service_profile: ServiceStackProfile | str,
-    node_provider_request: NodeProviderSelectionRequest | None,
-    configuration_validation: ConfigurationValidationService | None = None,
-    project_paths: ProjectPaths | None = None,
-    allow_wsl_windows_filesystem: bool = False,
-) -> PreflightService:
-    paths = project_paths or default_project_paths()
-    configuration = _preflight_configuration_for_provider(service_profile, node_provider_request)
-    evaluator, authorizer = _build_project_filesystem_services()
-    return PreflightService(
-        HostPreflightProbe(
-            project_paths=paths,
-            host_environment_detector=build_host_environment_detector(),
-            process_runner=build_process_runner(),
-        ),
-        replace(configuration, required_ports=()),
-        configuration_validation=configuration_validation,
-        project_filesystem_evaluator=evaluator,
-        project_filesystem_authorizer=authorizer,
-        project_path=paths.repository_root.as_posix(),
-        allow_wsl_windows_filesystem=allow_wsl_windows_filesystem,
-    )
-
-
-def _lxc_backend_for_provider_request(
-    provider_request: NodeProviderSelectionRequest,
-) -> ManagedLxcBackend | None:
-    if provider_request.requested_provider != NodeProviderKind.LXC_NATIVE:
-        return None
-    if provider_request.preferred_backend is not None:
-        if provider_request.preferred_backend in _LXC_SUPPORTED_BACKENDS:
-            return provider_request.preferred_backend
-        return None
-    for backend in provider_request.backend_candidates:
-        if backend not in _LXC_SUPPORTED_BACKENDS:
+_BOUNDARY_MODULES = {
+    "build_host_preparation_service": "composition_platform",
+    "build_platform_services": "composition_platform",
+    "build_artifact_services_for_provider": "composition_artifacts",
+    "build_lxc_artifact_services": "composition_artifacts",
+    "build_deployment_services_for_provider": "composition_deployment",
+    "build_lxc_deployment_services": "composition_deployment",
+    "build_setup_services": "composition_setup",
+    "build_application_services": "composition_setup",
+    "run_setup_with_terminal_status": "composition_setup",
+}
+
+
+def _delegate(name: str, *args: Any, **kwargs: Any) -> Any:
+    _sync_compatibility_overrides()
+    module_name = _BOUNDARY_MODULES.get(name)
+    if module_name is None:
+        return getattr(_runtime, name)(*args, **kwargs)
+    module = importlib.import_module(f".{module_name}", __package__)
+    return getattr(module, name)(*args, **kwargs)
+
+
+def _sync_compatibility_overrides() -> None:
+    """Make legacy facade patch points visible to the internal runtime module.
+
+    Existing tests and integrations patch selected composition symbols. The
+    facade remains the supported patch/import surface, so an explicitly
+    replaced facade attribute is copied to the runtime module immediately
+    before a delegated builder runs. Unmodified builder wrappers restore the
+    original runtime callable and never create a recursive alias.
+    """
+
+    for name, original in _RUNTIME_ORIGINALS.items():
+        facade_value = globals().get(name, original)
+        if facade_value is _FACADE_DEFAULTS[name]:
+            setattr(_runtime, name, original)
+        else:
+            setattr(_runtime, name, facade_value)
+
+    for name, original in _RUNTIME_DEFAULTS.items():
+        if name in _RUNTIME_BUILDER_NAMES:
             continue
-        if shutil.which(backend_cli(backend)):
-            return backend
-    return None
+        setattr(_runtime, name, globals().get(name, original))
+
+    for module_name in set(_BOUNDARY_MODULES.values()):
+        module = sys.modules.get(f"{__package__}.{module_name}")
+        if module is None:
+            continue
+        refresh = getattr(module, "_refresh_runtime_symbols", None)
+        if refresh is not None:
+            refresh()
+        boundary_defaults = getattr(module, "_BOUNDARY_DEFAULTS", {})
+        for name, default in boundary_defaults.items():
+            facade_value = globals().get(name)
+            if facade_value is _FACADE_DEFAULTS.get(name, default):
+                setattr(module, name, default)
+            else:
+                setattr(module, name, facade_value)
 
 
-def _default_node_provider_request() -> NodeProviderSelectionRequest:
-    provider_config = NodeProviderConfigYamlRepository(
-        project_paths=default_project_paths()
-    ).load()
-    return _node_provider_request_from_config(provider_config)
+def __getattr__(name: str) -> Any:
+    """Expose established runtime symbols without copying implementation here."""
 
+    value = getattr(_runtime, name)
+    if isinstance(value, FunctionType):
+        def delegated(*args: Any, **kwargs: Any) -> Any:
+            return _delegate(name, *args, **kwargs)
 
-def _node_provider_request_from_config(
-    provider_config: NodeProviderConfig,
-) -> NodeProviderSelectionRequest:
-    return NodeProviderSelectionRequest(
-        requested_provider=provider_config.default_provider,
-        preferred_backend=provider_config.preferred_backend,
-        backend_candidates=provider_config.backend_candidates,
-    )
-
-
-def _preflight_configuration_for_provider(
-    service_profile: ServiceStackProfile | str,
-    node_provider_request: NodeProviderSelectionRequest | None,
-) -> PreflightConfiguration:
-    configuration = replace(
-        default_preflight_configuration(service_profile=service_profile),
-        windows_wsl_bridge_required=_windows_wsl_bridge_required(),
-    )
-    profile_name = (
-        service_profile.value
-        if isinstance(service_profile, ServiceStackProfile)
-        else str(service_profile)
-    )
-    profiles = default_resource_profiles()
-    resource_profile = profiles.get(profile_name, profiles["default"])
-    configuration = replace(
-        configuration,
-        resources=ResourceThresholds(
-            minimum_cpu_count=resource_profile.minimum.cpu_threads,
-            minimum_memory_bytes=resource_profile.minimum.memory_bytes,
-            minimum_disk_free_bytes=resource_profile.minimum.free_disk_bytes,
-            disk_path=configuration.resources.disk_path,
-        ),
-    )
-    provider_request = node_provider_request or _default_node_provider_request()
-    if provider_request.requested_provider is not NodeProviderKind.LXC_NATIVE:
-        return replace(
-            configuration,
-            provider_metadata=ProviderPreflightMetadata(
-                provider=provider_request.requested_provider.value,
-                generic_checks=configuration.provider_metadata.generic_checks,
-            ),
-        )
-    backend = _lxc_backend_for_provider_request(provider_request)
-    if backend is None:
-        if not provider_request.backend_candidates:
-            raise ValueError(
-                "LXC-native preflight requires at least one managed backend candidate."
-            )
-        provider_dependencies = tuple(
-            RequiredDependency(backend_cli(candidate))
-            for candidate in provider_request.backend_candidates
-        )
-        backend_label = "auto"
-        provider_checks = tuple(
-            f"backend-cli:{backend_cli(candidate)}"
-            for candidate in provider_request.backend_candidates
-        )
-    else:
-        provider_dependencies = (RequiredDependency(backend_cli(backend)),)
-        backend_label = backend.value
-        provider_checks = (f"backend-cli:{backend_cli(backend)}",)
-    return replace(
-        configuration,
-        required_dependencies=(
-            *configuration.required_dependencies,
-            *provider_dependencies,
-        ),
-        provider_metadata=ProviderPreflightMetadata(
-            provider=provider_request.requested_provider.value,
-            backend=backend_label,
-            generic_checks=configuration.provider_metadata.generic_checks,
-            provider_checks=provider_checks,
-            daemon_checks=(
-                "managed-lxc-daemon-selected-backend",
-            ),
-            network_checks=(
-                "selected-backend-control-network",
-            ),
-            resource_expectations=(
-                "selected-backend-storage-pool",
-                "docker-swarm-profile",
-            ),
-        ),
-    )
-
-
-def _windows_wsl_bridge_required() -> bool:
-    value = os.environ.get(WINDOWS_EXPOSURE_ENVIRONMENT, "").strip().casefold()
-    return value not in {"0", "false", "no", "off", "disabled"}
-
-
-def _platform_init_steps(
-    *,
-    provider_request: NodeProviderSelectionRequest,
-    node_provider_selection: NodeProviderSelectionService,
-) -> tuple[AsyncWorkflowStep, ...]:
-    node_steps = tuple(
-        NodeProviderEnsureNodeStep(node, node_provider_selection, provider_request)
-        for node in DEFAULT_LXC_PLATFORM_NODES
-    )
-    return node_steps
-
-
-def _cluster_docker_steps(
-    lxc_docker_install: LxcDockerInstallService,
-) -> tuple[AsyncWorkflowStep, ...]:
-    return (LxcDockerInstallStep(lxc_docker_install, DEFAULT_LXC_PLATFORM_NODES),)
-
-
-def _cluster_swarm_bootstrap_steps(
-    lxc_swarm_bootstrap: LxcSwarmBootstrapService,
-) -> tuple[AsyncWorkflowStep, ...]:
-    return (
-        LxcSwarmBootstrapStep(lxc_swarm_bootstrap, DEFAULT_LXC_PLATFORM_NODES),
-    )
-
-
-def _platform_reconcile_steps(
-    *,
-    provider_request: NodeProviderSelectionRequest,
-    node_provider_selection: NodeProviderSelectionService,
-) -> tuple[AsyncWorkflowStep, ...]:
-    return tuple(
-        NodeProviderEnsureNodeStep(node, node_provider_selection, provider_request)
-        for node in DEFAULT_LXC_PLATFORM_NODES
-    )
-
-
-def _platform_expose_steps(
-    lxc_service_exposure: LxcServiceExposureService,
-    socat_manager: SocatManager,
-    socat_exposure: PortWslSocatExposure,
-    *,
-    service_profile: ServiceStackProfile,
-    live_consent: LiveConsent | None,
-) -> tuple[AsyncWorkflowStep, ...]:
-    return (
-        LxcServiceExposureStep(lxc_service_exposure),
-        _WslSocatExposeStep(
-            socat_manager,
-            socat_exposure,
-            service_profile=ServiceStackProfile(service_profile),
-            live_consent=live_consent,
-        ),
-    )
-
-
-def _platform_repair_lxc_proxy_drift_steps(
-    lxc_proxy_drift_repair: LxcProxyDriftRepairService,
-) -> tuple[AsyncWorkflowStep, ...]:
-    return (LxcProxyDriftRepairStep(lxc_proxy_drift_repair),)
-
-
-def _platform_verify_steps(
-    post_install_preflight: PreflightService,
-    *,
-    provider_request: NodeProviderSelectionRequest,
-    node_provider_selection: NodeProviderSelectionService,
-    lxc_service_exposure: LxcServiceExposureService,
-) -> tuple[AsyncWorkflowStep, ...]:
-    node_steps = tuple(
-        NodeProviderVerifyNodeStep(node, node_provider_selection, provider_request)
-        for node in DEFAULT_LXC_PLATFORM_NODES
-    )
-    return (
-        post_install_preflight,
-        *node_steps,
-        LxcServiceExposureVerifyStep(lxc_service_exposure),
-        _portainer_endpoint_verify_step(provider_request),
-    )
-
-
-def _cluster_verify_steps(
-    lxc_docker_verify: LxcDockerInstallService,
-    lxc_swarm_verify: LxcSwarmBootstrapService,
-) -> tuple[AsyncWorkflowStep, ...]:
-    return (
-        LxcDockerVerifyStep(lxc_docker_verify, DEFAULT_LXC_PLATFORM_NODES),
-        LxcSwarmVerifyStep(lxc_swarm_verify, DEFAULT_LXC_PLATFORM_NODES),
-    )
-
-
-def _portainer_endpoint_verify_step(
-    provider_request: NodeProviderSelectionRequest,
-) -> AsyncWorkflowStep:
-    backend = _lxc_backend_for_provider_request(provider_request)
-    if backend is None:
-        return _BlockedPlatformProviderStep(
-            target_id=PortainerEndpointVerifyStep.verification_target_id,
-            provider_request=provider_request,
-            message="Portainer endpoint verification requires a selected LXC backend.",
-            reason=LXC_BACKEND_REQUIRED_REASON,
-        )
-    return PortainerEndpointVerifyStep(
-        EnsurePortainerEndpoint(
-            portainer_client=LxcPortainerHttpClient(
-                backend=backend,
-                username="admin",
-                password=_operator_secret_value("TSW_PORTAINER_ADMIN_PASSWORD"),
-                stack_request_timeout_seconds=_operator_config_int(
-                    PORTAINER_STACK_REQUEST_TIMEOUT_ENVIRONMENT,
-                    DEFAULT_PORTAINER_STACK_REQUEST_TIMEOUT_SECONDS,
-                    minimum=1,
-                ),
-            ),
-            endpoint_name=DEFAULT_PORTAINER_ENDPOINT_NAME,
-            max_attempts=1,
-            wait_seconds=0,
-        )
-    )
-
-
-def _wsl_socat_forwarding_plans(
-    service_profile: ServiceStackProfile,
-) -> tuple[PortForwardingPlan, ...]:
-    return tuple(
-        PortForwardingPlan(
-            strategy=ForwardingStrategy.WSL2_SOCAT,
-            service=requirement.service,
-            listen_port=requirement.port,
-            target_port=requirement.port,
-            remediation=("Start WSL socat forwarding after live consent.",),
-        )
-        for requirement in default_setup_manifest(
-            service_profile=service_profile
-        ).required_ports
-    )
-
-
-def _wsl_socat_expose_message(status: VerificationStatus) -> str:
-    if status == VerificationStatus.VERIFIED:
-        return "WSL port exposure is configured for published service ports."
-    return "WSL port exposure failed for one or more published service ports."
-
-
-def _platform_reset_steps(
-    provider_request: NodeProviderSelectionRequest,
-    node_provider_selection: NodeProviderSelectionService,
-) -> tuple[AsyncWorkflowStep, ...]:
-    return (
-        NodeProviderResetManagedNodesStep(
-            DEFAULT_LXC_PLATFORM_NODES,
-            node_provider_selection,
-            provider_request,
-        ),
-    )
-
-
-def _platform_destroy_steps(
-    provider_request: NodeProviderSelectionRequest,
-    node_provider_selection: NodeProviderSelectionService,
-) -> tuple[AsyncWorkflowStep, ...]:
-    return (
-        NodeProviderDestroyManagedNodesStep(
-            DEFAULT_LXC_PLATFORM_NODES,
-            node_provider_selection,
-            provider_request,
-        ),
-    )
-
-
-async def _platform_init_pre_apply_guard(
-    preflight: PreflightService,
-    node_provider_selection: NodeProviderSelectionService,
-    provider_request: NodeProviderSelectionRequest,
-    live_consent: LiveConsent | None,
-) -> object:
-    preflight_result = await preflight.run(live_consent)
-    if not preflight_result.passed:
-        return preflight_result
-    return await node_provider_selection.verify_provider_selection(provider_request)
-
-
-def _operator_secret_value(name: str) -> str:
-    return os.environ.get(name) or f"<operator-supplied:{name}>"
-
-
-def _endpoint_status(
-    session: requests.Session,
-    url: str,
-    *,
-    timeout_seconds: int,
-) -> str:
-    try:
-        response = session.get(
-            url,
-            timeout=timeout_seconds,
-            allow_redirects=False,
-        )
-    except requests.Timeout:
-        return "timeout"
-    except requests.RequestException:
-        return "connection_error"
-    return f"http_{response.status_code}"
-
-
-def _endpoint_status_ready(status: str) -> bool:
-    if not status.startswith("http_"):
-        return False
-    try:
-        status_code = int(status.removeprefix("http_"))
-    except ValueError:
-        return False
-    return 100 <= status_code < 500
-
-
-def _endpoint_readiness_evidence(
-    service_stack: ServiceStackContract,
-    statuses: dict[str, str],
-    *,
-    phase: str,
-    attempt: int,
-) -> dict[str, str]:
-    return {
-        "attempt": str(attempt),
-        "endpoint_statuses": ",".join(
-            f"{name}={status}" for name, status in sorted(statuses.items())
-        ),
-        "evidence_kind": "service_endpoint_http",
-        "phase": phase,
-        "stack_name": service_stack.stack_name,
-    }
-
-
-def _wsl_lxc_lifecycle_capability_available() -> bool:
-    return (
-        _wsl_unprivileged_userns_clone_available()
-        and Path("/sys/fs/cgroup/cgroup.controllers").exists()
-        and Path("/proc/self/uid_map").exists()
-    )
-
-
-def _wsl_unprivileged_userns_clone_available(
-    path: Path = Path("/proc/sys/kernel/unprivileged_userns_clone"),
-) -> bool:
-    if not path.exists():
-        return True
-    return _linux_text_file_equals(path, "1")
-
-
-def _linux_text_file_equals(path: Path, expected: str) -> bool:
-    try:
-        return path.read_text(encoding="utf-8", errors="ignore").strip() == expected
-    except OSError:
-        return False
-
-
-def _operator_config_value(name: str, default: str) -> str:
-    return os.environ.get(name) or default
-
-
-def _secret_mode() -> str:
-    mode = _operator_config_value(SECRETS_MODE_ENVIRONMENT, "generated").strip()
-    if mode not in SECRET_MODES:
-        raise ValueError("TSW_SECRETS_MODE must be one of generated, fixed, or infisical.")
-    return mode
-
-
-def _fixed_secret_env_file() -> Path:
-    return Path(
-        _operator_config_value(
-            FIXED_SECRET_ENV_FILE_ENVIRONMENT,
-            DEFAULT_FIXED_SECRET_ENV_FILE.as_posix(),
-        )
-    )
-
-
-def _operator_config_int(name: str, default: int, *, minimum: int) -> int:
-    raw_value = os.environ.get(name, "").strip()
-    if not raw_value:
-        return default
-    try:
-        value = int(raw_value, 10)
-    except ValueError as exc:
-        raise ValueError(f"{name} must be an integer.") from exc
-    if value < minimum:
-        raise ValueError(f"{name} must be at least {minimum}.")
+        delegated.__name__ = value.__name__
+        delegated.__doc__ = value.__doc__
+        return delegated
     return value
 
 
-def _operator_config_float(name: str, default: float, *, minimum: float) -> float:
-    raw_value = os.environ.get(name, "").strip()
-    if not raw_value:
-        return default
-    try:
-        value = float(raw_value)
-    except ValueError as exc:
-        raise ValueError(f"{name} must be a number.") from exc
-    if value < minimum:
-        raise ValueError(f"{name} must be at least {minimum:g}.")
-    return value
+def __dir__() -> list[str]:
+    return sorted(set(globals()) | set(dir(_runtime)))
 
 
-def _lxc_proxy_listen_address() -> str:
-    address = _operator_config_value(
-        LXC_PROXY_LISTEN_ADDRESS_ENVIRONMENT,
-        DEFAULT_LXC_PROXY_LISTEN_ADDRESS,
-    ).strip()
-    if address not in {"127.0.0.1", "0.0.0.0"}:
-        raise ValueError("LXC proxy listen address must be 127.0.0.1 or 0.0.0.0.")
-    return address
+def build_application_logger(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_application_logger", *args, **kwargs)
 
 
-def _lxc_docker_registry_mirror_configuration() -> DockerRegistryMirrorConfiguration | None:
-    mirror_url = os.getenv("TSW_LXC_DOCKER_REGISTRY_MIRROR", "").strip()
-    if not mirror_url:
-        return None
-    return DockerRegistryMirrorConfiguration(mirror_url)
+def build_host_environment_detector(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_host_environment_detector", *args, **kwargs)
 
 
-def _lxc_docker_apt_mirror_configuration() -> DockerAptMirrorConfiguration | None:
-    configuration = DockerAptMirrorConfiguration(
-        ubuntu_archive_url=os.getenv(LXC_UBUNTU_APT_MIRROR_ENVIRONMENT, "").strip()
-        or None,
-        ubuntu_security_url=os.getenv(LXC_UBUNTU_SECURITY_APT_MIRROR_ENVIRONMENT, "").strip()
-        or None,
-        docker_apt_url=os.getenv(LXC_DOCKER_APT_MIRROR_ENVIRONMENT, "").strip()
-        or None,
-        docker_gpg_url=os.getenv(LXC_DOCKER_APT_GPG_URL_ENVIRONMENT, "").strip()
-        or None,
-    )
-    if not configuration.configured:
-        return None
-    return configuration
+def build_host_detection_service(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_host_detection_service", *args, **kwargs)
 
 
-def _lxc_reachable_host_ip() -> str:
-    for interface_name in ("incusbr0",):
-        address = _host_ipv4_for_interface(interface_name)
-        if address:
-            return address
-    return ""
+def build_host_preparation_service(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_host_preparation_service", *args, **kwargs)
 
 
-def _host_ipv4_for_interface(interface_name: str) -> str:
-    try:
-        result = subprocess.run(
-            [
-                "ip",
-                "-4",
-                "-o",
-                "addr",
-                "show",
-                "dev",
-                interface_name,
-            ],
-            capture_output=True,
-            text=True,
-            check=False,
-            shell=False,
-            timeout=10,
-        )
-    except (OSError, subprocess.TimeoutExpired):
-        return ""
-    if result.returncode != 0:
-        return ""
-    match = re.search(r"\binet\s+(?P<address>\d+\.\d+\.\d+\.\d+)/", result.stdout)
-    return match.group("address") if match else ""
+def build_project_filesystem_inspector(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_project_filesystem_inspector", *args, **kwargs)
 
 
-def _lxc_manager_node() -> NodeSpec:
-    manager = next(
-        (node for node in DEFAULT_LXC_PLATFORM_NODES if node.role == NodeRole.MANAGER),
-        None,
-    )
-    if manager is None:
-        raise ValueError("LXC platform node list must include a manager.")
-    return manager
+def build_preflight_service(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_preflight_service", *args, **kwargs)
 
 
-def _operator_config_source_ref(name: str) -> str:
-    if os.environ.get(name):
-        return "operator_env"
-    return "default"
+def build_configuration_validation_service(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_configuration_validation_service", *args, **kwargs)
 
 
-def _add_optional_config(environment: dict[str, str], name: str) -> None:
-    value = os.environ.get(name, "").strip()
-    if value:
-        environment[name] = value
+def build_compose_file_repository(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_compose_file_repository", *args, **kwargs)
 
 
-def _deployment_stack_environment(
-    service_profile: ServiceStackProfile,
-) -> dict[str, dict[str, str]]:
-    registry_endpoint = _swarm_registry_endpoint()
-    environment = {
-        "traefik": {
-            TRAEFIK_IMAGE_ENVIRONMENT: _operator_config_value(
-                TRAEFIK_IMAGE_ENVIRONMENT,
-                "traefik:v3.7.4",
-            ),
-            TRAEFIK_TLS_CERT_SECRET_NAME_ENVIRONMENT: _operator_config_value(
-                TRAEFIK_TLS_CERT_SECRET_NAME_ENVIRONMENT,
-                DEFAULT_TRAEFIK_TLS_CERT_SECRET_NAME,
-            ),
-            TRAEFIK_TLS_KEY_SECRET_NAME_ENVIRONMENT: _operator_config_value(
-                TRAEFIK_TLS_KEY_SECRET_NAME_ENVIRONMENT,
-                DEFAULT_TRAEFIK_TLS_KEY_SECRET_NAME,
-            ),
-        },
-        "nexus": {
-            NEXUS_IMAGE_ENVIRONMENT: _operator_config_value(
-                NEXUS_IMAGE_ENVIRONMENT,
-                "sonatype/nexus3:3.75.1",
-            ),
-        },
-        "jenkins": {
-            JENKINS_IMAGE_ENVIRONMENT: _operator_config_value(
-                JENKINS_IMAGE_ENVIRONMENT,
-                f"{registry_endpoint}/jenkins:0.2.0",
-            ),
-            "TSW_JENKINS_ADMIN_PASSWORD": _operator_secret_value("TSW_JENKINS_ADMIN_PASSWORD"),
-        },
-        "pulsar": {
-            PULSAR_IMAGE_ENVIRONMENT: _operator_config_value(
-                PULSAR_IMAGE_ENVIRONMENT,
-                DEFAULT_PULSAR_IMAGE,
-            ),
-            PULSAR_MANAGER_IMAGE_ENVIRONMENT: _operator_config_value(
-                PULSAR_MANAGER_IMAGE_ENVIRONMENT,
-                DEFAULT_PULSAR_MANAGER_IMAGE,
-            ),
-            PULSAR_MANAGER_BOOTSTRAP_IMAGE_ENVIRONMENT: _operator_config_value(
-                PULSAR_MANAGER_BOOTSTRAP_IMAGE_ENVIRONMENT,
-                "python:3.12.13-alpine3.23",
-            ),
-            "TSW_PULSAR_TOKEN_SECRET_KEY": _operator_secret_value("TSW_PULSAR_TOKEN_SECRET_KEY"),
-            "TSW_PULSAR_ADMIN_TOKEN": _operator_secret_value("TSW_PULSAR_ADMIN_TOKEN"),
-            "TSW_PULSAR_MANAGER_ADMIN_PASSWORD": _operator_secret_value("TSW_PULSAR_MANAGER_ADMIN_PASSWORD"),
-        },
-        "sonarqube": {
-            "TSW_SONARQUBE_POSTGRES_PASSWORD": _operator_secret_value("TSW_SONARQUBE_POSTGRES_PASSWORD"),
-            "TSW_POSTGRES_PASSWORD": _operator_secret_value("TSW_POSTGRES_PASSWORD"),
-        }
-    }
-    if service_profile is not ServiceStackProfile.SERVICE_ACCESS:
-        return environment
-
-    environment["service-access"] = {
-        SERVICE_ACCESS_DASHBOARD_IMAGE_ENVIRONMENT: _operator_config_value(
-            SERVICE_ACCESS_DASHBOARD_IMAGE_ENVIRONMENT,
-            f"{registry_endpoint}/service-access-dashboard:0.2.0",
-        ),
-        SERVICE_ACCESS_NGINX_IMAGE_ENVIRONMENT: _operator_config_value(
-            SERVICE_ACCESS_NGINX_IMAGE_ENVIRONMENT,
-            f"{registry_endpoint}/service-access-nginx:0.2.0",
-        ),
-    }
-    environment["infisical"] = {
-        INFISICAL_ENCRYPTION_KEY_ENVIRONMENT: _operator_secret_value(
-            INFISICAL_ENCRYPTION_KEY_ENVIRONMENT,
-        ),
-        INFISICAL_AUTH_SECRET_ENVIRONMENT: _operator_secret_value(
-            INFISICAL_AUTH_SECRET_ENVIRONMENT,
-        ),
-        INFISICAL_LOGIN_EMAIL_ENVIRONMENT: _operator_secret_value(
-            INFISICAL_LOGIN_EMAIL_ENVIRONMENT,
-        ),
-        INFISICAL_PASSWORD_ENVIRONMENT: _operator_secret_value(
-            INFISICAL_PASSWORD_ENVIRONMENT,
-        ),
-        INFISICAL_ADMIN_FIRST_NAME_ENVIRONMENT: _operator_config_value(
-            INFISICAL_ADMIN_FIRST_NAME_ENVIRONMENT,
-            "Tiny",
-        ),
-        INFISICAL_ADMIN_LAST_NAME_ENVIRONMENT: _operator_config_value(
-            INFISICAL_ADMIN_LAST_NAME_ENVIRONMENT,
-            "Admin",
-        ),
-        INFISICAL_POSTGRES_PASSWORD_ENVIRONMENT: _operator_secret_value(
-            INFISICAL_POSTGRES_PASSWORD_ENVIRONMENT,
-        ),
-        INFISICAL_REDIS_PASSWORD_ENVIRONMENT: _operator_secret_value(
-            INFISICAL_REDIS_PASSWORD_ENVIRONMENT,
-        ),
-    }
-    _add_optional_config(
-        environment["infisical"],
-        INFISICAL_IMAGE_ENVIRONMENT,
-    )
-    _add_optional_config(
-        environment["infisical"],
-        INFISICAL_POSTGRES_IMAGE_ENVIRONMENT,
-    )
-    _add_optional_config(
-        environment["infisical"],
-        INFISICAL_REDIS_IMAGE_ENVIRONMENT,
-    )
-    return environment
+def build_network_doctor_service(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_network_doctor_service", *args, **kwargs)
 
 
-def _container_image_contracts_from_environment() -> tuple[ContainerImageContract, ...]:
-    return resolve_container_image_contracts(DEFAULT_CONTAINER_IMAGE_CONTRACTS, os.environ)
+def build_read_only_hang_diagnostics(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_read_only_hang_diagnostics", *args, **kwargs)
 
 
-def _split_image_ref(image_ref: str) -> tuple[str, str]:
-    if "@" in image_ref:
-        image_name, digest = image_ref.rsplit("@", 1)
-        return image_name, f"@{digest}"
-    if ":" not in image_ref.rsplit("/", 1)[-1]:
-        return image_ref, "latest"
-    image_name, tag = image_ref.rsplit(":", 1)
-    return image_name, tag
+def build_preflight_evidence_writer(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_preflight_evidence_writer", *args, **kwargs)
 
 
-def _nexus_docker_hub_proxy_repository_name() -> str:
-    repository_name = _operator_config_value(
-        NEXUS_DOCKER_HUB_PROXY_REPOSITORY_ENVIRONMENT,
-        DEFAULT_NEXUS_DOCKER_HUB_PROXY_REPOSITORY,
-    ).strip()
-    if not repository_name:
-        raise ValueError("Nexus Docker Hub proxy repository name must not be empty.")
-    return repository_name
+def build_process_runner(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_process_runner", *args, **kwargs)
 
 
-def _nexus_docker_hub_proxy_port() -> int:
-    raw_port = _operator_config_value(
-        NEXUS_DOCKER_HUB_PROXY_PORT_ENVIRONMENT,
-        str(DEFAULT_NEXUS_DOCKER_HUB_PROXY_PORT),
-    ).strip()
-    try:
-        port = int(raw_port)
-    except ValueError as exc:
-        raise ValueError("Nexus Docker Hub proxy port must be an integer.") from exc
-    if port <= 0 or port > 65535:
-        raise ValueError("Nexus Docker Hub proxy port must be a valid TCP port.")
-    return port
+def build_network_repair_service(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_network_repair_service", *args, **kwargs)
 
 
-def _nexus_docker_proxy_remote_url() -> str:
-    mirror_configuration = _lxc_docker_registry_mirror_configuration()
-    remote_url = (
-        mirror_configuration.mirror_url
-        if mirror_configuration is not None
-        else DEFAULT_NEXUS_DOCKER_HUB_PROXY_REMOTE_URL
-    )
-    if not _is_http_url(remote_url):
-        raise ValueError("Nexus Docker proxy remote URL must be HTTP or HTTPS.")
-    return remote_url
+def build_network_repair_options(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_network_repair_options", *args, **kwargs)
 
 
-def _swarm_registry_endpoint() -> str:
-    endpoint = _operator_config_value(
-        SWARM_REGISTRY_ENDPOINT_ENVIRONMENT,
-        DEFAULT_SWARM_REGISTRY_ENDPOINT,
-    ).strip()
-    if not REGISTRY_ENDPOINT_PATTERN.fullmatch(endpoint):
-        raise ValueError(
-            "Swarm registry endpoint must be host[:port] without scheme or credentials."
-        )
-    return endpoint
+def build_post_install_preflight_service(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_post_install_preflight_service", *args, **kwargs)
 
 
-def _is_http_url(value: str) -> bool:
-    parsed = urlparse(value)
-    return parsed.scheme in {"http", "https"} and bool(parsed.netloc)
+def build_setup_ui(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_setup_ui", *args, **kwargs)
 
 
-def _local_http_url(host: str, port: str) -> str:
-    scheme = "http"
-    return f"{scheme}://{host}:{port}"
+async def run_setup_with_terminal_status(*args: Any, **kwargs: Any) -> Any:
+    return await _delegate("run_setup_with_terminal_status", *args, **kwargs)
 
 
-def _infisical_secret_seed_steps(
-    service_profile: ServiceStackProfile,
-) -> list[EnsureInfisicalSecretItems]:
-    if service_profile is not ServiceStackProfile.SERVICE_ACCESS:
-        return []
-    if os.environ.get(SEED_INFISICAL_ITEMS_ENVIRONMENT) != "1":
-        return []
-    return [
-        EnsureInfisicalSecretItems(
-            infisical_client=PlaywrightInfisicalClient(
-                base_url=_operator_config_value(
-                    INFISICAL_URL_ENVIRONMENT,
-                    "https://localhost",
-                ),
-            ),
-            login_email=_operator_secret_value(INFISICAL_LOGIN_EMAIL_ENVIRONMENT),
-            password=_operator_secret_value(INFISICAL_PASSWORD_ENVIRONMENT),
-            items=_infisical_seed_items(),
-        ),
-    ]
+def build_platform_services(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_platform_services", *args, **kwargs)
 
 
-def _infisical_apply_readiness_steps(
-    service_profile: ServiceStackProfile,
-    *,
-    service_stack_by_name: dict[str, ServiceStackContract],
-) -> tuple[EndpointReadinessCheck, ...]:
-    if service_profile is not ServiceStackProfile.SERVICE_ACCESS:
-        readiness_steps: list[EndpointReadinessCheck] = []
-        return tuple(readiness_steps)
-    attempts = _operator_config_int(
-        INFISICAL_READINESS_ATTEMPTS_ENVIRONMENT,
-        DEFAULT_INFISICAL_READINESS_ATTEMPTS,
-        minimum=1,
-    )
-    interval = _operator_config_float(
-        INFISICAL_READINESS_INTERVAL_ENVIRONMENT,
-        DEFAULT_INFISICAL_READINESS_INTERVAL_SECONDS,
-        minimum=0,
-    )
-    readiness_steps = [
-        EndpointReadinessCheck(
-            service_stack_by_name["infisical"],
-            verification_target_id="deployment:infisical-bootstrap-service-readiness",
-            max_attempts=attempts,
-            wait_seconds=int(interval),
-        ),
-        EndpointReadinessCheck(
-            service_stack_by_name["service-access"],
-            verification_target_id="deployment:infisical-bootstrap-access-readiness",
-            max_attempts=attempts,
-            wait_seconds=int(interval),
-        ),
-    ]
-    return tuple(readiness_steps)
+def build_artifact_services_for_provider(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_artifact_services_for_provider", *args, **kwargs)
 
 
-def _infisical_bootstrap_steps(
-    service_profile: ServiceStackProfile,
-    *,
-    cli: InfisicalCliClient | None = None,
-    swarm_runtime: LxcSwarmRuntime | None = None,
-) -> list[EnsureInfisicalSilentInstall]:
-    if service_profile is not ServiceStackProfile.SERVICE_ACCESS:
-        return []
-    return [
-        EnsureInfisicalSilentInstall(
-            cli=cli or InfisicalCliClient(),
-            storage=LocalFileStorage(),
-            bootstrap_client=InfisicalBootstrapHttpClient(
-                base_url=_operator_config_value(
-                    INFISICAL_URL_ENVIRONMENT,
-                    _local_http_url("localhost", "17080"),
-                ),
-                readiness_attempts=_operator_config_int(
-                    INFISICAL_READINESS_ATTEMPTS_ENVIRONMENT,
-                    DEFAULT_INFISICAL_READINESS_ATTEMPTS,
-                    minimum=1,
-                ),
-                readiness_interval_seconds=_operator_config_float(
-                    INFISICAL_READINESS_INTERVAL_ENVIRONMENT,
-                    DEFAULT_INFISICAL_READINESS_INTERVAL_SECONDS,
-                    minimum=0,
-                ),
-                readiness_recovery=(
-                    swarm_runtime.recover_infisical_migration_lock
-                    if swarm_runtime is not None
-                    else None
-                ),
-            ),
-            config=InfisicalSilentInstallConfig(
-                external_url=_operator_config_value(
-                    INFISICAL_URL_ENVIRONMENT,
-                    _local_http_url("localhost", "17080"),
-                ),
-                internal_url=_operator_config_value(
-                    INFISICAL_INTERNAL_URL_ENVIRONMENT,
-                    _local_http_url("infisical", "8080"),
-                ),
-                admin_email=_required_operator_secret_value(
-                    INFISICAL_LOGIN_EMAIL_ENVIRONMENT,
-                ),
-                admin_first_name=_operator_config_value(
-                    INFISICAL_ADMIN_FIRST_NAME_ENVIRONMENT,
-                    "Tiny",
-                ),
-                admin_last_name=_operator_config_value(
-                    INFISICAL_ADMIN_LAST_NAME_ENVIRONMENT,
-                    "Admin",
-                ),
-                admin_password=_required_operator_secret_value(
-                    INFISICAL_PASSWORD_ENVIRONMENT,
-                ),
-                organization=_operator_config_value(
-                    INFISICAL_ORGANIZATION_ENVIRONMENT,
-                    DEFAULT_INFISICAL_ORGANIZATION,
-                ),
-                encryption_key=_required_operator_secret_value(
-                    INFISICAL_ENCRYPTION_KEY_ENVIRONMENT,
-                ),
-                auth_secret=_required_operator_secret_value(
-                    INFISICAL_AUTH_SECRET_ENVIRONMENT,
-                ),
-                postgres_password=_required_operator_secret_value(
-                    INFISICAL_POSTGRES_PASSWORD_ENVIRONMENT,
-                ),
-                redis_password=_operator_secret_value(INFISICAL_REDIS_PASSWORD_ENVIRONMENT),
-            ),
-        ),
-    ]
+def build_lxc_artifact_services(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_lxc_artifact_services", *args, **kwargs)
 
 
-def _with_infisical_post_apply_steps(
-    application_steps: tuple[object, ...],
-    post_steps: tuple[object, ...],
-) -> tuple[object, ...]:
-    if not post_steps:
-        return application_steps
-    ordered_steps: list[object] = []
-    inserted = False
-    for step in application_steps:
-        ordered_steps.append(step)
-        service_stack = getattr(step, "service_stack", None)
-        if getattr(service_stack, "stack_name", "") == "service-access":
-            ordered_steps.extend(post_steps)
-            inserted = True
-    if not inserted:
-        ordered_steps.extend(post_steps)
-    return tuple(ordered_steps)
+def build_deployment_services_for_provider(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_deployment_services_for_provider", *args, **kwargs)
 
 
-def _prioritize_infisical_apply_steps(
-    application_steps: tuple[object, ...],
-) -> tuple[object, ...]:
-    priority_stack_names = ("traefik", "infisical", "service-access")
-    prioritized: list[object] = []
-    remaining = list(application_steps)
-    for stack_name in priority_stack_names:
-        for index, step in enumerate(remaining):
-            service_stack = getattr(step, "service_stack", None)
-            if getattr(service_stack, "stack_name", "") == stack_name:
-                prioritized.append(step)
-                del remaining[index]
-                break
-    return (*prioritized, *remaining)
+def build_lxc_deployment_services(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_lxc_deployment_services", *args, **kwargs)
 
 
-def _with_post_stack_steps(
-    application_steps: tuple[object, ...],
-    stack_name: str,
-    post_steps: tuple[object, ...],
-) -> tuple[object, ...]:
-    if not post_steps:
-        return application_steps
-    ordered_steps: list[object] = []
-    inserted = False
-    for step in application_steps:
-        ordered_steps.append(step)
-        service_stack = getattr(step, "service_stack", None)
-        if getattr(service_stack, "stack_name", "") == stack_name:
-            ordered_steps.extend(post_steps)
-            inserted = True
-    if not inserted:
-        ordered_steps.extend(post_steps)
-    return tuple(ordered_steps)
+def build_setup_services(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_setup_services", *args, **kwargs)
 
 
-def _infisical_seed_items() -> tuple[InfisicalSecretItem, ...]:
-    return (
-        InfisicalSecretItem(
-            "platform/jenkins",
-            _operator_config_value("TSW_JENKINS_ADMIN_USERNAME", "admin"),
-            _required_operator_secret_value("TSW_JENKINS_ADMIN_PASSWORD"),
-        ),
-        InfisicalSecretItem(
-            "platform/nexus",
-            _operator_config_value("TSW_NEXUS_ADMIN_USERNAME", "admin"),
-            _required_operator_secret_value("TSW_NEXUS_ADMIN_PASSWORD"),
-        ),
-        InfisicalSecretItem(
-            "platform/portainer",
-            _operator_config_value("TSW_PORTAINER_USERNAME", "admin"),
-            _required_operator_secret_value("TSW_PORTAINER_ADMIN_PASSWORD"),
-        ),
-        InfisicalSecretItem(
-            "platform/pulsar",
-            "admin",
-            _required_operator_secret_value("TSW_PULSAR_ADMIN_TOKEN"),
-        ),
-        InfisicalSecretItem(
-            "platform/pulsar-manager",
-            "admin",
-            _required_operator_secret_value("TSW_PULSAR_MANAGER_ADMIN_PASSWORD"),
-        ),
-        InfisicalSecretItem(
-            "platform/sonarqube",
-            _operator_config_value("TSW_SONARQUBE_ADMIN_USERNAME", "admin"),
-            _required_operator_secret_value("TSW_SONARQUBE_ADMIN_PASSWORD"),
-        ),
-    )
+def build_application_services(*args: Any, **kwargs: Any) -> Any:
+    return _delegate("build_application_services", *args, **kwargs)
 
 
-def _required_operator_secret_value(name: str) -> str:
-    value = os.environ.get(name, "").strip()
-    if not value:
-        raise ValueError(f"Required operator secret is missing: {name}")
-    return value
+_FACADE_DEFAULTS = {
+    name: globals()[name]
+    for name in _RUNTIME_BUILDER_NAMES
+}
+_RUNTIME_ORIGINALS = {
+    name: getattr(_runtime, name)
+    for name in _RUNTIME_BUILDER_NAMES
+}
+_RUNTIME_PATCHABLE_NAMES = frozenset(dir(_runtime))
+_RUNTIME_DEFAULTS = {
+    name: getattr(_runtime, name)
+    for name in _RUNTIME_PATCHABLE_NAMES
+}
+__all__ = [name for name in dir(_runtime) if not name.startswith("_")] + list(
+    _RUNTIME_BUILDER_NAMES
+)
